jaleco/cischeat.cpp and related drivers: Make road and sprite generator as device, Updates
Common:
- Reduce preprocessor defines
- Improve logging and save state supports
- Reduce runtime tag lookups and literal tag usage
- Make some variables constant
- Minor optimizations
- Use BIT helper for single bit values
- Fix namings
- Reduce duplicates
- Use shortened typename values for consistency
- Add notes

jaleco/cischeat.cpp:
- Consolidated driver into single file
- Split scudhamm and related drivers into separate driver files
