+++ src/mame/jaleco/cischeat.cpp
@@ -54,50 +54,28 @@ Year + Game     Hardware:   Main    Sub#1   Sub#2   Sound   Sound Chips
         MB - Middle board  (GFX)               GP-9189  EB90015-20038
         LB - Lower board   (CPU/GFX)           GP-9188A EB90015-20037-1
 
-[92 Arm Champs II]          68000   -       -       -       2xM6295
-
-    Top board   EB91042-20048-3 AC-91106B
-    Lower board EB91015-20038 GP-9189
-    Same chips as Scud Hammer
-
-[94 Scud Hammer]            68000   -       -       -       2xM6295
-
-    Board CF-92128B Chips:
-        GS9001501
-        GS90015-02 x 2  GS600406   x 2      [Tilemaps]
-        GS90015-03
-
-    Board GP-9189 Chips:
-        GS90015-03 x 2                      [Sprites]
-        GS90015-06 x 2
-        GS90015-07
-        GS90015-08
-        GS90015-09
-        GS90015-10
-        GS90015-11
-        MR90015-35 x 2
 ---------------------------------------------------------------------------
 
 ----------------------------------------------------------------------------------
-Main CPU                    [Cisco Heat]        [F1 GP Star]        [Scud Hammer]
+Main CPU                    [Cisco Heat]        [F1 GP Star]
 ----------------------------------------------------------------------------------
-ROM                 R       000000-07ffff       <                   <
-                            100000-17ffff       <                   RW : I / O + Sound
-Work RAM + Sprites  RW      0f0000-0fffff       <                   <
-Hardware Regs       RW      080000-087fff       <                   <
-Units Linking RAM   RW      088000-88ffff       <                   -
-Shared RAM #2       RW      090000-097fff       <                   -
-Shared RAM #1       RW      098000-09ffff       <                   -
-Scroll RAM 0        RW      0a0000-0a7fff       <                   <
-Scroll RAM 1        RW      0a8000-0affff       <                   -
-Scroll RAM 2        RW      0b0000-0b7fff       <                   <
-Palette RAM         RW      0b8000-0bffff       <                   <
-    Palette Scroll 0        0b9c00-0b9fff       0b9e00-0b9fff       <
-    Palette Scroll 1        0bac00-0bafff       0bae00-0bafff       -
-    Palette Road 0          0bb800-0bbfff       <                   -
-    Palette Road 1          0bc800-0bcfff       <                   -
-    Palette Sprites         0bd000-0bdfff       <                   0bb000-0bbfff
-    Palette Scroll 2        0bec00-0befff0      0bee00-0befff       0bce00-0bcfff
+ROM                 R       000000-07ffff       <
+                            100000-17ffff       <
+Work RAM + Sprites  RW      0f0000-0fffff       <
+Hardware Regs       RW      080000-087fff       <
+Units Linking RAM   RW      088000-88ffff       <
+Shared RAM #2       RW      090000-097fff       <
+Shared RAM #1       RW      098000-09ffff       <
+Scroll RAM 0        RW      0a0000-0a7fff       <
+Scroll RAM 1        RW      0a8000-0affff       <
+Scroll RAM 2        RW      0b0000-0b7fff       <
+Palette RAM         RW      0b8000-0bffff       <
+    Palette Scroll 0        0b9c00-0b9fff       0b9e00-0b9fff
+    Palette Scroll 1        0bac00-0bafff       0bae00-0bafff
+    Palette Road 0          0bb800-0bbfff       <
+    Palette Road 1          0bc800-0bcfff       <
+    Palette Sprites         0bd000-0bdfff       <
+    Palette Scroll 2        0bec00-0befff0      0bee00-0befff
 ----------------------------------------------------------------------------------
 
 ----------------------------------------------------------------
@@ -149,21 +127,14 @@ Common Issues:
 - Screen control register (priorities, layers enabling etc.) - Where is it?
 - Sound communication not quite right: see Test Mode
 
-To Do:
+TODO:
 
 - Priorities!
-- Use the Tilemap Manager for the road layers (when this kind of layers
-  will be supported) for performance and better priority support.
-  A line based zooming is additionally needed for f1gpstar.
 - Wild Pilot road needs some serious work (empty gaps, ship "floating"
   in stage 6 etc.);
 - Force feedback :)
 - Major cleanups needed, especially in video file;
-- scudhamm: during attract mode every other how to play screen shows up with
-  wrong background layer colors;
-- armchmp2: arm input device not completely understood;
-- Split the non-road games into own driver, merge them with Alien Command,
-  device-ify the sprite chip;
+- Verify sprite lag frames
 
 BTANBs:
 - In cischeat & bigrun, at the start of some levels, you can see the empty
@@ -182,18 +153,517 @@ Cisco Heat.
 ******************************************************************************/
 
 #include "emu.h"
-#include "cischeat.h"
+
+#include "jaleco_road.h"
+#include "jaleco_zoomspr.h"
+#include "ms1_gatearray.h"
+#include "ms1_tmap.h"
 
 #include "cpu/m68000/m68000.h"
-#include "machine/adc0804.h"
-#include "machine/nvram.h"
+#include "machine/gen_latch.h"
+#include "machine/ticket.h"
+#include "machine/timer.h"
+#include "machine/watchdog.h"
+
 #include "sound/okim6295.h"
 #include "sound/ymopm.h"
+
+#include "emupal.h"
+#include "screen.h"
 #include "speaker.h"
+
+#include <algorithm>
 
 #include "cischeat.lh"
 #include "f1gpstar.lh"
-#include "captflag.lh"
+
+
+#define LOG_UNKNOWN (1 << 1)
+
+#define VERBOSE (0)
+
+#include "logmacro.h"
+
+#define LOGUNKNOWN(...)   LOGMASKED(LOG_UNKNOWN, __VA_ARGS__) 
+
+namespace {
+// TODO: better inheritance, eventually split individual driver files
+
+class cischeat_state : public driver_device
+{
+public:
+	cischeat_state(const machine_config &mconfig, device_type type, const char *tag)
+		: driver_device(mconfig, type, tag)
+		, m_maincpu(*this, "maincpu")
+		, m_road1cpu(*this, "road1cpu")
+		, m_road2cpu(*this, "road2cpu")
+		, m_iocpu(*this, "iocpu")
+		, m_soundcpu(*this, "soundcpu")
+		, m_screen(*this, "screen")
+		, m_watchdog(*this, "watchdog")
+		, m_oki(*this, "oki%u", 1U)
+		, m_palette(*this, "palette")
+		, m_soundlatch(*this, "soundlatch%u", 1U)
+		, m_gatearray(*this, "gatearray")
+		, m_spritegen(*this, "spritegen")
+		, m_road(*this, "road%u", 0U)
+		, m_tmap(*this, "scroll%u", 0U)
+		, m_ram(*this, "ram")
+		, m_roadram(*this, "road%u", 0U)
+		, m_f1gpstr2_ioready(*this, "ioready")
+		, m_io_in(*this, "IN%u", 0U)
+		, m_io_pedal(*this, "PEDAL")
+		, m_leds(*this, "led%u", 0U)
+	{}
+
+	void init_cischeat();
+	void init_bigrun();
+	void init_f1gpstar();
+
+	void cischeat(machine_config &config);
+	void f1gpstr2(machine_config &config);
+	void f1gpstar(machine_config &config);
+	void bigrun(machine_config &config);
+	void bigrun_d65006(machine_config &config);
+
+protected:
+	virtual void machine_start() override ATTR_COLD;
+	virtual void video_start() override ATTR_COLD;
+
+	// TODO: make these to have a more meaningful name
+	required_device<cpu_device> m_maincpu;
+	required_device<cpu_device> m_road1cpu;
+	required_device<cpu_device> m_road2cpu;
+	optional_device<cpu_device> m_iocpu;
+	required_device<cpu_device> m_soundcpu;
+	required_device<screen_device> m_screen;
+	optional_device<watchdog_timer_device> m_watchdog;
+	required_device_array<okim6295_device, 2> m_oki;
+	required_device<palette_device> m_palette;
+	required_device_array<generic_latch_16_device, 2> m_soundlatch;
+	optional_device<megasys1_gatearray_device> m_gatearray;
+	required_device<jaleco_zoomspr_device> m_spritegen;
+	required_device_array<jaleco_road_device, 2> m_road;
+	required_device_array<megasys1_tilemap_device, 3> m_tmap;
+
+	required_shared_ptr<u16> m_ram;
+	required_shared_ptr_array<u16, 2> m_roadram;
+	optional_shared_ptr<u16> m_f1gpstr2_ioready;
+
+	optional_ioport_array<7> m_io_in;
+	optional_ioport m_io_pedal;
+
+	output_finder<3> m_leds;
+
+	u16 *m_spriteram = nullptr;
+	u16 m_active_layers = 0U;
+
+	s32 m_ip_select = 0;
+
+	u8 m_motor_value = 0U;
+	u8 m_io_value = 0U;
+
+	void bigrun_soundbank_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	u16 bigrun_ip_select_r();
+	void leds_out_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	void unknown_out_w(u16 data);
+	void motor_out_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	void wheel_out_w(u16 data);
+	void ip_select_w(u16 data);
+	void ip_select_plus1_w(u16 data);
+	void bigrun_comms_w(u16 data);
+	void active_layers_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	u16 cischeat_ip_select_r();
+	void cischeat_soundlatch_w(u16 data);
+	void cischeat_comms_w(u16 data);
+	u16 f1gpstar_wheel_r();
+	u16 f1gpstr2_ioready_r();
+	void f1gpstar_motor_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	void f1gpstar_soundint_w(u16 data);
+	void f1gpstar_comms_w(u16 data);
+	void f1gpstr2_io_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	template <unsigned Which> void cischeat_soundbank_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	void sound_irq(int state);
+	u32 screen_update_bigrun(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
+	u32 screen_update_cischeat(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
+	u32 screen_update_f1gpstar(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect);
+	TIMER_DEVICE_CALLBACK_MEMBER(bigrun_scanline);
+	void cischeat_untangle_sprites(const char *region);
+
+	void bigrun_map(address_map &map) ATTR_COLD;
+	void bigrun_map2(address_map &map) ATTR_COLD;
+	void bigrun_map3(address_map &map) ATTR_COLD;
+	void bigrun_sound_map(address_map &map) ATTR_COLD;
+	void cischeat_map(address_map &map) ATTR_COLD;
+	void cischeat_map2(address_map &map) ATTR_COLD;
+	void cischeat_map3(address_map &map) ATTR_COLD;
+	void cischeat_sound_map(address_map &map) ATTR_COLD;
+	void f1gpstar_map(address_map &map) ATTR_COLD;
+	void f1gpstar_map2(address_map &map) ATTR_COLD;
+	void f1gpstar_map3(address_map &map) ATTR_COLD;
+	void f1gpstar_sound_map(address_map &map) ATTR_COLD;
+	void f1gpstr2_io_map(address_map &map) ATTR_COLD;
+	void f1gpstr2_map(address_map &map) ATTR_COLD;
+	void f1gpstr2_sound_map(address_map &map) ATTR_COLD;
+};
+
+class wildplt_state : public cischeat_state
+{
+public:
+	wildplt_state(const machine_config &mconfig, device_type type, const char *tag)
+		: cischeat_state(mconfig, type, tag)
+		, m_io_x(*this, "P%uX", 1U)
+		, m_io_y(*this, "P%uY", 1U)
+		, m_io_in1_sub(*this, "IN1_%u", 1U)
+		, m_io_in1_common(*this, "IN1_COMMON")
+	{}
+
+	void wildplt(machine_config &config);
+
+protected:
+	virtual void machine_start() override ATTR_COLD;
+	virtual void video_start() override ATTR_COLD;
+
+private:
+	required_ioport_array<2> m_io_x;
+	required_ioport_array<2> m_io_y;
+	required_ioport_array<2> m_io_in1_sub;
+	required_ioport m_io_in1_common;
+
+	u16 *m_buffer_spriteram = nullptr;
+	std::unique_ptr<u16[]> m_allocated_spriteram;
+
+	u16 m_sprite_dma_reg = 0U;
+
+	u16 m_wildplt_output = 0U;
+
+	u16 wildplt_xy_r();
+	u16 wildplt_mux_r();
+	void wildplt_mux_w(u16 data);
+
+	void sprite_dma_w(offs_t offset, u16 data, u16 mem_mask = ~0);
+	void wildplt_map(address_map &map) ATTR_COLD;
+};
+
+/***************************************************************************
+
+
+                            Video Hardware Init
+
+
+***************************************************************************/
+
+void cischeat_state::video_start()
+{
+	m_spriteram = &m_ram[0x8000/2];
+
+	m_active_layers = 0;
+}
+
+void wildplt_state::video_start()
+{
+	cischeat_state::video_start();
+	m_buffer_spriteram = &m_ram[0x8000 / 2];
+	m_allocated_spriteram = std::make_unique<u16[]>(0x1000 / 2);
+	m_spriteram = m_allocated_spriteram.get();
+
+	save_item(NAME(m_sprite_dma_reg));
+	save_pointer(NAME(m_allocated_spriteram), 0x1000 / 2);
+}
+
+void wildplt_state::sprite_dma_w(offs_t offset, u16 data, u16 mem_mask)
+{
+	// bit 13: 0 -> 1 transition triggers a sprite DMA
+	if (data & 0x2000 && (m_sprite_dma_reg & 0x2000) == 0)
+	{
+		std::copy_n(&m_buffer_spriteram[0], 0x1000 / 2, &m_allocated_spriteram[0]);
+	}
+
+	// other bits unknown
+	COMBINE_DATA(&m_sprite_dma_reg);
+}
+
+
+/***************************************************************************
+
+
+                                Screen Drawing
+
+
+***************************************************************************/
+
+/**************************************************************************
+                                Big Run
+**************************************************************************/
+
+u32 cischeat_state::screen_update_bigrun(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect)
+{
+	bitmap.fill(0x1000, cliprect);
+
+	for (int i = 7; i >= 4; i--)
+	{
+		m_road[0]->draw(bitmap, cliprect, i, i, (i != 7));
+		m_road[1]->draw(bitmap, cliprect, i, i, true);
+		m_spritegen->draw_sprites(bitmap, cliprect, i + 1, i, m_spriteram, 0x1000 / 2);
+	}
+
+	m_tmap[0]->draw(screen, bitmap, cliprect, 0, 0);
+	m_tmap[1]->draw(screen, bitmap, cliprect, 0, 0);
+
+	for (int i = 3; i >= 0; i--)
+	{
+		m_road[0]->draw(bitmap, cliprect, i, i, true);
+		m_road[1]->draw(bitmap, cliprect, i, i, true);
+		m_spritegen->draw_sprites(bitmap, cliprect, i + 1, i, m_spriteram, 0x1000 / 2);
+	}
+
+	m_tmap[2]->draw(screen, bitmap, cliprect, 0, 0);
+
+	return 0;
+}
+
+
+/**************************************************************************
+                                Cisco Heat
+**************************************************************************/
+
+u32 cischeat_state::screen_update_cischeat(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect)
+{
+	bitmap.fill(0, cliprect);
+
+	m_road[0]->draw(bitmap, cliprect, 7, 5, false);
+	m_road[1]->draw(bitmap, cliprect, 7, 5, true);
+
+	m_tmap[0]->draw(screen, bitmap, cliprect, 0, 0);
+	m_tmap[1]->draw(screen, bitmap, cliprect, 0, 0);
+
+	m_spritegen->draw_sprites(bitmap, cliprect, 15, 3, m_spriteram, 0x1000 / 2);
+	m_road[0]->draw(bitmap, cliprect, 4, 1, true);
+	m_road[1]->draw(bitmap, cliprect, 4, 1, true);
+	m_spritegen->draw_sprites(bitmap, cliprect, 2, 2, m_spriteram, 0x1000 / 2);
+	m_road[0]->draw(bitmap, cliprect, 0, 0, true);
+	m_road[1]->draw(bitmap, cliprect, 0, 0, true);
+	m_spritegen->draw_sprites(bitmap, cliprect, 1, 0, m_spriteram, 0x1000 / 2);
+	m_tmap[2]->draw(screen, bitmap, cliprect, 0, 0);
+
+	/* for the map screen */
+	m_spritegen->draw_sprites(bitmap, cliprect, 16, 16, m_spriteram, 0x1000 / 2);
+
+	return 0;
+}
+
+
+/**************************************************************************
+                            F1 GrandPrix Star
+**************************************************************************/
+
+u32 cischeat_state::screen_update_f1gpstar(screen_device &screen, bitmap_ind16 &bitmap, const rectangle &cliprect)
+{
+	bitmap.fill(0, cliprect);
+
+/*  1: clouds 5, grad 7, road 0     2: clouds 5, grad 7, road 0, tunnel roof 0 */
+
+	/* road 1!! 0!! */
+	m_road[1]->draw(bitmap, cliprect, 6, 7, false);
+	m_road[0]->draw(bitmap, cliprect, 6, 7, true);
+
+	m_tmap[0]->draw(screen, bitmap, cliprect, 0, 0);
+	m_tmap[1]->draw(screen, bitmap, cliprect, 0, 0);
+
+	/* road 1!! 0!! */
+	m_road[1]->draw(bitmap, cliprect, 1, 5, true);
+	m_road[0]->draw(bitmap, cliprect, 1, 5, true);
+
+	m_spritegen->draw_sprites(bitmap, cliprect, 15, 2, m_spriteram, 0x1000 / 2);
+
+	/* road 1!! 0!! */
+	m_road[1]->draw(bitmap, cliprect, 0, 0, true);
+	m_road[0]->draw(bitmap, cliprect, 0, 0, true);
+
+	m_spritegen->draw_sprites(bitmap, cliprect, 1, 1, m_spriteram, 0x1000 / 2);
+	m_tmap[2]->draw(screen, bitmap, cliprect, 0, 0);
+	m_spritegen->draw_sprites(bitmap, cliprect, 0, 0, m_spriteram, 0x1000 / 2);
+
+	return 0;
+}
+
+/***************************************************************************
+
+
+                        Hardware registers access
+
+
+***************************************************************************/
+
+/*
+    F1 GP Star has a real pedal, while Cisco Heat's is connected to
+    a switch. The Former game stores, during boot, the value that
+    corresponds to the pedal not pressed, and compares against it:
+
+    The value returned must decrease when the pedal is pressed.
+*/
+
+/**************************************************************************
+                                Big Run
+**************************************************************************/
+
+u16 cischeat_state::bigrun_ip_select_r()
+{
+	switch (m_ip_select & 0x3)
+	{
+		case 0 : return m_io_in[6]->read();      // Driving Wheel
+		case 1 : return 0xffff;                 // Cockpit: Up / Down Position
+		case 2 : return 0xffff;                 // Cockpit: Left / Right Position?
+		case 3 : return m_io_pedal->read();    // Accelerator (Pedal)
+		default: return 0xffff;
+	}
+}
+
+
+void cischeat_state::unknown_out_w(u16 data)
+{
+	// ?? 91/1/91/1 ...
+	LOGUNKNOWN("%s: Unknown unknown_out_w write %04x", machine().describe_context(), data);
+}
+
+
+void cischeat_state::motor_out_w(offs_t offset, u16 data, u16 mem_mask)
+{
+	// motor (seat?)
+	if (ACCESSING_BITS_0_7)
+		m_leds[2] = (data & 0xff) != m_motor_value ? 1 : 0;
+	m_motor_value = data & 0xff;
+}
+
+
+void cischeat_state::wheel_out_w(u16 data)
+{
+	// motor (wheel?)
+	LOGUNKNOWN("%s: Unknown wheel_out_w write %04x", machine().describe_context(), data);
+}
+
+
+void cischeat_state::ip_select_w(u16 data)
+{
+	m_ip_select = data;
+}
+
+
+void cischeat_state::ip_select_plus1_w(u16 data)
+{
+	// value above + 1
+	m_ip_select = data + 1;
+}
+
+
+void cischeat_state::bigrun_comms_w(u16 data)
+{
+	/* Not sure about this one.. */
+	m_road1cpu->set_input_line(INPUT_LINE_RESET, BIT(data, 1) ? ASSERT_LINE : CLEAR_LINE);
+	m_road2cpu->set_input_line(INPUT_LINE_RESET, BIT(data, 1) ? ASSERT_LINE : CLEAR_LINE);
+	m_soundcpu->set_input_line(INPUT_LINE_RESET, BIT(data, 0) ? ASSERT_LINE : CLEAR_LINE);
+}
+
+// TODO: fake port, never written to my knowledge!
+void cischeat_state::active_layers_w(offs_t offset, u16 data, u16 mem_mask)
+{
+	COMBINE_DATA(&m_active_layers);
+}
+
+
+/**************************************************************************
+                                Cisco Heat
+**************************************************************************/
+
+u16 cischeat_state::cischeat_ip_select_r()
+{
+	switch (m_ip_select & 0x3)
+	{
+		case 0 : return m_io_in[6]->read();  // Driving Wheel
+		case 1 : return ~0;                 // Cockpit: Up / Down Position?
+		case 2 : return ~0;                 // Cockpit: Left / Right Position?
+		default: return ~0;
+	}
+}
+
+
+void cischeat_state::cischeat_soundlatch_w(u16 data)
+{
+	/* Sound CPU: reads latch during int 4, and stores command */
+	m_soundlatch[0]->write(data);
+	m_soundcpu->set_input_line(4, HOLD_LINE);
+}
+
+
+void cischeat_state::cischeat_comms_w(u16 data)
+{
+	/* Not sure about this one.. */
+	m_road1cpu->set_input_line(INPUT_LINE_RESET, BIT(data, 1) ? ASSERT_LINE : CLEAR_LINE);
+	m_road2cpu->set_input_line(INPUT_LINE_RESET, BIT(data, 1) ? ASSERT_LINE : CLEAR_LINE);
+	m_soundcpu->set_input_line(INPUT_LINE_RESET, BIT(data, 0) ? ASSERT_LINE : CLEAR_LINE);
+}
+
+
+
+/**************************************************************************
+                            F1 GrandPrix Star
+**************************************************************************/
+
+u16 cischeat_state::f1gpstar_wheel_r()
+{
+	return (m_io_pedal->read() & 0xff) + ((m_io_in[5]->read() & 0xff) << 8);
+}
+
+
+u16 cischeat_state::f1gpstr2_ioready_r()
+{
+	return BIT(m_f1gpstr2_ioready[0], 0) ? 0xff : 0xf0;
+}
+
+
+void cischeat_state::f1gpstar_motor_w(offs_t offset, u16 data, u16 mem_mask)
+{
+	// "shudder" motors, leds
+	if (ACCESSING_BITS_0_7)
+	{
+		machine().bookkeeping().coin_counter_w(0, BIT(data, 0));
+		machine().bookkeeping().coin_counter_w(1, BIT(data, 1));
+		m_leds[0] = BIT(data, 2);   // start button
+		m_leds[1] = BIT(data, 5);   // ?
+		// wheel | seat motor
+		m_leds[2] = BIT(data, 3) | BIT(data, 4);
+	}
+}
+
+
+void cischeat_state::f1gpstar_soundint_w(u16 data)
+{
+	/* $80008 and $80018 usually written in sequence, but not always */
+	m_soundcpu->set_input_line(4, HOLD_LINE);
+}
+
+
+void cischeat_state::f1gpstar_comms_w(u16 data)
+{
+	/* Not sure about this one. Values: $10 then 0, $7 then 0 */
+	m_road1cpu->set_input_line(INPUT_LINE_RESET, BIT(data, 0) ? ASSERT_LINE : CLEAR_LINE);
+	m_road2cpu->set_input_line(INPUT_LINE_RESET, BIT(data, 1) ? ASSERT_LINE : CLEAR_LINE);
+	m_soundcpu->set_input_line(INPUT_LINE_RESET, BIT(data, 2) ? ASSERT_LINE : CLEAR_LINE);
+}
+
+
+void cischeat_state::f1gpstr2_io_w(offs_t offset, u16 data, u16 mem_mask)
+{
+	if (ACCESSING_BITS_0_7)
+	{
+		if ((m_io_value & 4) && ((data & 4) == 0))
+			m_iocpu->set_input_line(4, HOLD_LINE);
+		if ((m_io_value & 2) && ((data & 2) == 0))
+			m_iocpu->set_input_line(2, HOLD_LINE);
+		m_io_value = data & 0xff;
+	}
+}
+
 
 /**************************************************************************
 
@@ -203,18 +673,28 @@ Cisco Heat.
 
 **************************************************************************/
 
+void cischeat_state::machine_start()
+{
+	m_leds.resolve();
+	m_motor_value = 0;
+	m_io_value = 0;
+
+	save_item(NAME(m_ip_select));
+	save_item(NAME(m_motor_value));
+	save_item(NAME(m_io_value));
+}
 
 /**************************************************************************
                                 Big Run
 **************************************************************************/
 
-void cischeat_state::leds_out_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void cischeat_state::leds_out_w(offs_t offset, u16 data, u16 mem_mask)
 {
 	// leds
 	if (ACCESSING_BITS_0_7)
 	{
-		machine().bookkeeping().coin_counter_w(0, data & 0x01);
-		machine().bookkeeping().coin_counter_w(1, data & 0x02);
+		machine().bookkeeping().coin_counter_w(0, BIT(data, 0));
+		machine().bookkeeping().coin_counter_w(1, BIT(data, 1));
 		m_leds[0] = BIT(data, 4);   // start button
 		m_leds[1] = BIT(data, 5);   // ?
 	}
@@ -227,14 +707,14 @@ void cischeat_state::bigrun_map(address_map &map)
 	map(0x080002, 0x080003).portr("IN2").w(FUNC(cischeat_state::unknown_out_w));    // Buttons
 	map(0x080004, 0x080005).portr("IN3").w(FUNC(cischeat_state::motor_out_w));      // Motor Limit Switches
 	map(0x080006, 0x080007).portr("IN4").w(FUNC(cischeat_state::wheel_out_w));      // DSW 1 & 2
-	map(0x080008, 0x080009).r(m_soundlatch2, FUNC(generic_latch_16_device::read));  // From sound cpu
-	map(0x08000a, 0x08000b).w(m_soundlatch, FUNC(generic_latch_16_device::write));  // To sound cpu
+	map(0x080008, 0x080009).r(m_soundlatch[1], FUNC(generic_latch_16_device::read));  // From sound cpu
+	map(0x08000a, 0x08000b).w(m_soundlatch[0], FUNC(generic_latch_16_device::write));  // To sound cpu
 	map(0x08000c, 0x08000d).nopw();            // ??
 	map(0x080010, 0x080011).rw(FUNC(cischeat_state::bigrun_ip_select_r), FUNC(cischeat_state::ip_select_w));
 	map(0x080012, 0x080013).w(FUNC(cischeat_state::ip_select_plus1_w));
-	map(0x082000, 0x082005).rw("scroll0", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).rw("scroll1", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082100, 0x082105).rw("scroll2", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082000, 0x082005).rw(m_tmap[0], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082008, 0x08200d).rw(m_tmap[1], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082100, 0x082105).rw(m_tmap[2], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
 	map(0x082108, 0x082109).noprw();                    // ? written with 0 only
 	map(0x082200, 0x082201).portr("IN5");               // DSW 3 (4 bits)
 	map(0x082208, 0x082209).noprw();                    // watchdog reset
@@ -246,17 +726,17 @@ void cischeat_state::bigrun_map(address_map &map)
 	    Amusingly, if you run the communication test as ID = X then soft reset -> ID = Y, what was at ID = X gets an OK in the second test
 	    so it's likely to be the only thing needed. */
 	map(0x084000, 0x0847ff).ram();                                                 // Linking with other units
-	map(0x088000, 0x08bfff).ram().share("share2"); // Sharedram with sub CPU#2
-	map(0x08c000, 0x08ffff).ram().share("share1"); // Sharedram with sub CPU#1
+	map(0x088000, 0x08bfff).ram().share("road2_shareram"); // Sharedram with sub CPU#2
+	map(0x08c000, 0x08ffff).ram().share("road1_shareram"); // Sharedram with sub CPU#1
 
 	/* Only writes to the first 0x40000 bytes affect the tilemaps:             */
 	/* either these games support larger tilemaps or have more ram than needed */
-	map(0x090000, 0x093fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0");  // Scroll ram 0
-	map(0x094000, 0x097fff).ram().w("scroll1", FUNC(megasys1_tilemap_device::write)).share("scroll1");  // Scroll ram 1
-	map(0x098000, 0x09bfff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2");  // Scroll ram 2
+	map(0x090000, 0x093fff).ram().w(m_tmap[0], FUNC(megasys1_tilemap_device::write)).share("scroll0");  // Scroll ram 0
+	map(0x094000, 0x097fff).ram().w(m_tmap[1], FUNC(megasys1_tilemap_device::write)).share("scroll1");  // Scroll ram 1
+	map(0x098000, 0x09bfff).ram().w(m_tmap[2], FUNC(megasys1_tilemap_device::write)).share("scroll2");  // Scroll ram 2
 
 	map(0x09c000, 0x09ffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");         // Palettes
-	map(0x0f0000, 0x0fffff).ram().share("ram");                                                         // RAM
+	map(0x0f0000, 0x0fffff).ram().share(m_ram);                                                         // RAM
 	map(0x100000, 0x13ffff).rom().region("user1", 0);                                                   // ROM
 }
 
@@ -282,17 +762,17 @@ void cischeat_state::cischeat_map(address_map &map)
 	map(0x080002, 0x080003).portr("IN2").w(FUNC(cischeat_state::unknown_out_w));    // Buttons
 	map(0x080004, 0x080005).portr("IN3").w(FUNC(cischeat_state::motor_out_w));      // Motor Limit Switches
 	map(0x080006, 0x080007).portr("IN4").w(FUNC(cischeat_state::wheel_out_w));      // DSW 1 & 2
-	map(0x08000a, 0x08000b).w(m_soundlatch, FUNC(generic_latch_16_device::write));  // To sound cpu
+	map(0x08000a, 0x08000b).w(m_soundlatch[0], FUNC(generic_latch_16_device::write));  // To sound cpu
 	map(0x08000c, 0x08000d).nopw();            // ??
 	map(0x080010, 0x080011).rw(FUNC(cischeat_state::cischeat_ip_select_r), FUNC(cischeat_state::ip_select_w));
 	map(0x080012, 0x080013).nopw();            // value above + 1
-	map(0x082000, 0x082005).rw("scroll0", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).rw("scroll1", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082100, 0x082105).rw("scroll2", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082000, 0x082005).rw(m_tmap[0], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082008, 0x08200d).rw(m_tmap[1], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082100, 0x082105).rw(m_tmap[2], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
 	map(0x082108, 0x082109).noprw();                 // ? written with 0 only
 	map(0x082200, 0x082201).portr("IN5");    // DSW 3 (4 bits)
 	map(0x082208, 0x082209).noprw();                 // watchdog reset
-	map(0x082300, 0x082301).r(m_soundlatch2, FUNC(generic_latch_16_device::read)).w(FUNC(cischeat_state::cischeat_soundlatch_w)); // From sound cpu
+	map(0x082300, 0x082301).r(m_soundlatch[1], FUNC(generic_latch_16_device::read)).w(FUNC(cischeat_state::cischeat_soundlatch_w)); // From sound cpu
 	map(0x082308, 0x082309).w(FUNC(cischeat_state::cischeat_comms_w));
 	map(0x082400, 0x082401).w(FUNC(cischeat_state::active_layers_w));
 
@@ -304,17 +784,17 @@ void cischeat_state::cischeat_map(address_map &map)
     No mem access error from the other CPU's, though.. */
 
 	/* this is the right order of sharedram's */
-	map(0x090000, 0x097fff).ram().share("share2"); // Sharedram with sub CPU#2
-	map(0x098000, 0x09ffff).ram().share("share1"); // Sharedram with sub CPU#1
+	map(0x090000, 0x097fff).ram().share("road2_shareram"); // Sharedram with sub CPU#2
+	map(0x098000, 0x09ffff).ram().share("road1_shareram"); // Sharedram with sub CPU#1
 
 	/* Only writes to the first 0x40000 bytes affect the tilemaps:             */
 	/* either these games support larger tilemaps or have more ram than needed */
-	map(0x0a0000, 0x0a7fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0");  // Scroll ram 0
-	map(0x0a8000, 0x0affff).ram().w("scroll1", FUNC(megasys1_tilemap_device::write)).share("scroll1");  // Scroll ram 1
-	map(0x0b0000, 0x0b7fff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2");  // Scroll ram 2
+	map(0x0a0000, 0x0a7fff).ram().w(m_tmap[0], FUNC(megasys1_tilemap_device::write)).share("scroll0");  // Scroll ram 0
+	map(0x0a8000, 0x0affff).ram().w(m_tmap[1], FUNC(megasys1_tilemap_device::write)).share("scroll1");  // Scroll ram 1
+	map(0x0b0000, 0x0b7fff).ram().w(m_tmap[2], FUNC(megasys1_tilemap_device::write)).share("scroll2");  // Scroll ram 2
 	map(0x0b8000, 0x0bffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");         // Palettes
 
-	map(0x0f0000, 0x0fffff).ram().share("ram");                                                         // RAM
+	map(0x0f0000, 0x0fffff).ram().share(m_ram);                                                         // RAM
 	map(0x100000, 0x17ffff).rom().region("user1", 0);                                                   // ROM
 }
 
@@ -344,16 +824,16 @@ void cischeat_state::f1gpstar_map(address_map &map)
 	map(0x080000, 0x080001).portr("IN1");    // DSW 1 & 2
 	map(0x080004, 0x080005).portr("IN2").w(FUNC(cischeat_state::f1gpstar_motor_w));   // Buttons
 	map(0x080006, 0x080007).portr("IN3");    // ? Read at boot only
-	map(0x080008, 0x080009).r(m_soundlatch2, FUNC(generic_latch_16_device::read));     // From sound cpu
-	map(0x080008, 0x080009).w(m_soundlatch, FUNC(generic_latch_16_device::write));    // To sound cpu
+	map(0x080008, 0x080009).r(m_soundlatch[1], FUNC(generic_latch_16_device::read));     // From sound cpu
+	map(0x080008, 0x080009).w(m_soundlatch[0], FUNC(generic_latch_16_device::write));    // To sound cpu
 	map(0x08000c, 0x08000d).portr("IN4");    // DSW 3
 	map(0x080010, 0x080011).r(FUNC(cischeat_state::f1gpstar_wheel_r)).nopw(); // Accel + Driving Wheel
 	map(0x080014, 0x080015).nopw();
 	map(0x080018, 0x080019).w(FUNC(cischeat_state::f1gpstar_soundint_w));
 
-	map(0x082000, 0x082005).rw("scroll0", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).rw("scroll1", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082100, 0x082105).rw("scroll2", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082000, 0x082005).rw(m_tmap[0], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082008, 0x08200d).rw(m_tmap[1], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082100, 0x082105).rw(m_tmap[2], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
 	map(0x082108, 0x082109).noprw();                 // ? written with 0 only
 	map(0x082208, 0x082209).noprw();                 // watchdog reset
 	map(0x082308, 0x082309).nopr().w(FUNC(cischeat_state::f1gpstar_comms_w));
@@ -361,17 +841,17 @@ void cischeat_state::f1gpstar_map(address_map &map)
 
 	map(0x088000, 0x0883ff).ram();                                                                     // Linking with other units
 
-	map(0x090000, 0x097fff).ram().share("share2"); // Sharedram with sub CPU#2
-	map(0x098000, 0x09ffff).ram().share("share1"); // Sharedram with sub CPU#1
+	map(0x090000, 0x097fff).ram().share("road2_shareram"); // Sharedram with sub CPU#2
+	map(0x098000, 0x09ffff).ram().share("road1_shareram"); // Sharedram with sub CPU#1
 
 	/* Only writes to the first 0x40000 bytes affect the tilemaps:             */
 	/* either these games support larger tilemaps or have more ram than needed */
-	map(0x0a0000, 0x0a7fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0");     // Scroll ram 0
-	map(0x0a8000, 0x0affff).ram().w("scroll1", FUNC(megasys1_tilemap_device::write)).share("scroll1");     // Scroll ram 1
-	map(0x0b0000, 0x0b7fff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2");     // Scroll ram 2
+	map(0x0a0000, 0x0a7fff).ram().w(m_tmap[0], FUNC(megasys1_tilemap_device::write)).share("scroll0");     // Scroll ram 0
+	map(0x0a8000, 0x0affff).ram().w(m_tmap[1], FUNC(megasys1_tilemap_device::write)).share("scroll1");     // Scroll ram 1
+	map(0x0b0000, 0x0b7fff).ram().w(m_tmap[2], FUNC(megasys1_tilemap_device::write)).share("scroll2");     // Scroll ram 2
 	map(0x0b8000, 0x0bffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");              // Palettes
 
-	map(0x0f0000, 0x0fffff).ram().share("ram");                                             // RAM
+	map(0x0f0000, 0x0fffff).ram().share(m_ram);                                             // RAM
 	map(0x100000, 0x17ffff).rom().region("user1", 0);                                                            // ROM
 }
 
@@ -380,74 +860,81 @@ void cischeat_state::f1gpstar_map(address_map &map)
                             Wild Pilot
 **************************************************************************/
 
-// ad stick read select
-uint16_t cischeat_state::wildplt_xy_r()
+void wildplt_state::machine_start()
 {
-	switch(m_ip_select)
+	cischeat_state::machine_start();
+
+	save_item(NAME(m_wildplt_output));
+}
+
+// ad stick read select
+u16 wildplt_state::wildplt_xy_r()
+{
+	switch (m_ip_select)
 	{
-		case 1: return ioport("P2Y")->read() | (ioport("P2X")->read()<<8);
-		case 2: return ioport("P1Y")->read() | (ioport("P1X")->read()<<8);
+		case 1: return m_io_y[1]->read() | (m_io_x[1]->read() << 8);
+		case 2: return m_io_y[0]->read() | (m_io_x[0]->read() << 8);
 	}
 
 	return 0xffff;
 }
 
 // buttons & sensors are muxed. bit 0 routes to coin chute (single according to test mode)
-uint16_t cischeat_state::wildplt_mux_r()
+u16 wildplt_state::wildplt_mux_r()
 {
-	uint16_t split_in = 0xffff;
-	switch(m_wildplt_output & 0xc)
+	u16 split_in = 0xffff;
+	switch (m_wildplt_output & 0xc)
 	{
-//      case 0: return ioport("IN1")->read();
-		case 4: split_in = ioport("IN1_1")->read(); break;
-		case 8: split_in = ioport("IN1_2")->read(); break;
+//      case 0: return m_io_in[1]->read();
+		case 4: split_in = m_io_in1_sub[0]->read(); break;
+		case 8: split_in = m_io_in1_sub[1]->read(); break;
 	}
 
-	return split_in & ioport("IN1_COMMON")->read();
+	return split_in & m_io_in1_common->read();
 }
 
-void cischeat_state::wildplt_mux_w(uint16_t data)
+void wildplt_state::wildplt_mux_w(u16 data)
 {
 	m_wildplt_output = data & 0xc;
 }
 
 
-// Same as f1gpstar, but vregs are slightly different:
+// Same as f1gpstr2, but vregs are slightly different:
 void wildplt_state::wildplt_map(address_map &map)
 {
 	map(0x000000, 0x07ffff).rom();                                                                     // ROM
-	map(0x080000, 0x080001).portr("IN0").w(FUNC(cischeat_state::f1gpstr2_io_w));    // DSW 1 & 2
-	map(0x080004, 0x080005).r(FUNC(cischeat_state::wildplt_mux_r)).w(FUNC(cischeat_state::wildplt_mux_w)); // Buttons
-	map(0x080008, 0x080009).r(m_soundlatch2, FUNC(generic_latch_16_device::read));     // From sound cpu
-	map(0x080008, 0x080009).w(m_soundlatch, FUNC(generic_latch_16_device::write));    // To sound cpu
+	map(0x080000, 0x080001).portr("IN0").w(FUNC(wildplt_state::f1gpstr2_io_w));    // DSW 1 & 2
+	map(0x080004, 0x080005).r(FUNC(wildplt_state::wildplt_mux_r)).w(FUNC(wildplt_state::wildplt_mux_w)); // Buttons
+	map(0x080008, 0x080009).r(m_soundlatch[1], FUNC(generic_latch_16_device::read));     // From sound cpu
+	map(0x080008, 0x080009).w(m_soundlatch[0], FUNC(generic_latch_16_device::write));    // To sound cpu
 	map(0x08000c, 0x08000d).w(FUNC(wildplt_state::sprite_dma_w)); // 1000, 3000
-	map(0x080010, 0x080011).r(FUNC(cischeat_state::wildplt_xy_r)).w(FUNC(cischeat_state::ip_select_w)); // X, Y
+	map(0x080010, 0x080011).r(FUNC(wildplt_state::wildplt_xy_r)).w(FUNC(wildplt_state::ip_select_w)); // X, Y
 	map(0x080014, 0x080015).nopw();
-	map(0x080018, 0x080019).rw(FUNC(cischeat_state::f1gpstr2_ioready_r), FUNC(cischeat_state::f1gpstar_soundint_w));
+	map(0x080018, 0x080019).rw(FUNC(wildplt_state::f1gpstr2_ioready_r), FUNC(wildplt_state::f1gpstar_soundint_w));
 
-	map(0x081000, 0x081fff).ram().share("shareio");
+	map(0x081000, 0x081fff).ram().share("io_shareram");
 
-	map(0x082000, 0x082005).rw("scroll0", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).rw("scroll1", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082100, 0x082105).rw("scroll2", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082000, 0x082005).rw(m_tmap[0], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082008, 0x08200d).rw(m_tmap[1], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082100, 0x082105).rw(m_tmap[2], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
 	map(0x082108, 0x082109).noprw();                 // ? written with 0 only
 	map(0x082208, 0x082209).noprw();                 // watchdog reset
-	map(0x082308, 0x082309).nopr().w(FUNC(cischeat_state::f1gpstar_comms_w));
-	map(0x082400, 0x082401).w(FUNC(cischeat_state::active_layers_w));
+	map(0x082308, 0x082309).nopr().w(FUNC(wildplt_state::f1gpstar_comms_w));
+	map(0x082400, 0x082401).w(FUNC(wildplt_state::active_layers_w));
 
 //  map(0x088000, 0x088fff).ram(); // Linking with other units (not present on this)
 
-	map(0x090000, 0x097fff).ram().share("share2"); // Sharedram with sub CPU#2
-	map(0x098000, 0x09ffff).ram().share("share1"); // Sharedram with sub CPU#1
+	map(0x090000, 0x097fff).ram().share("road2_shareram"); // Sharedram with sub CPU#2
+	map(0x098000, 0x09ffff).ram().share("road1_shareram"); // Sharedram with sub CPU#1
 
 	/* Only writes to the first 0x40000 bytes affect the tilemaps:             */
 	/* either these games support larger tilemaps or have more ram than needed */
-	map(0x0a0000, 0x0a7fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0");     // Scroll ram 0
-	map(0x0a8000, 0x0affff).ram().w("scroll1", FUNC(megasys1_tilemap_device::write)).share("scroll1");     // Scroll ram 1
-	map(0x0b0000, 0x0b7fff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2");     // Scroll ram 2
+	map(0x0a0000, 0x0a7fff).ram().w(m_tmap[0], FUNC(megasys1_tilemap_device::write)).share("scroll0");     // Scroll ram 0
+	map(0x0a8000, 0x0affff).ram().w(m_tmap[1], FUNC(megasys1_tilemap_device::write)).share("scroll1");     // Scroll ram 1
+	map(0x0b0000, 0x0b7fff).ram().w(m_tmap[2], FUNC(megasys1_tilemap_device::write)).share("scroll2");     // Scroll ram 2
 	map(0x0b8000, 0x0bffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");               // Palettes
 
-	map(0x0f0000, 0x0fffff).ram().share("ram");                                             // RAM
+	map(0x0f0000, 0x0fffff).ram().share(m_ram);                                             // RAM
 	map(0x100000, 0x17ffff).rom().region("user1", 0);                                                            // ROM
 }
 
@@ -463,18 +950,18 @@ void cischeat_state::f1gpstr2_map(address_map &map)
 	map(0x080000, 0x080001).portr("IN1").w(FUNC(cischeat_state::f1gpstr2_io_w));      // DSW 1 & 2
 	map(0x080004, 0x080005).portr("IN2").w(FUNC(cischeat_state::f1gpstar_motor_w));   // Buttons
 	map(0x080006, 0x080007).portr("IN3");    // ? Read at boot only
-	map(0x080008, 0x080009).r(m_soundlatch2, FUNC(generic_latch_16_device::read));     // From sound cpu
-	map(0x080008, 0x080009).w(m_soundlatch, FUNC(generic_latch_16_device::write));    // To sound cpu
+	map(0x080008, 0x080009).r(m_soundlatch[1], FUNC(generic_latch_16_device::read));     // From sound cpu
+	map(0x080008, 0x080009).w(m_soundlatch[0], FUNC(generic_latch_16_device::write));    // To sound cpu
 	map(0x08000c, 0x08000d).portr("IN4");    // DSW 3
 	map(0x080010, 0x080011).r(FUNC(cischeat_state::f1gpstar_wheel_r)).nopw();
 	map(0x080014, 0x080015).nopw();
 	map(0x080018, 0x080019).rw(FUNC(cischeat_state::f1gpstr2_ioready_r), FUNC(cischeat_state::f1gpstar_soundint_w));
 
-	map(0x081000, 0x081fff).ram().share("shareio");
+	map(0x081000, 0x081fff).ram().share("io_shareram");
 
-	map(0x082000, 0x082005).rw("scroll0", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).rw("scroll1", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082100, 0x082105).rw("scroll2", FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082000, 0x082005).rw(m_tmap[0], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082008, 0x08200d).rw(m_tmap[1], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
+	map(0x082100, 0x082105).rw(m_tmap[2], FUNC(megasys1_tilemap_device::scroll_r), FUNC(megasys1_tilemap_device::scroll_w));
 	map(0x082108, 0x082109).noprw();                 // ? written with 0 only
 	map(0x082208, 0x082209).noprw();                 // watchdog reset
 	map(0x082308, 0x082309).nopr().w(FUNC(cischeat_state::f1gpstar_comms_w));
@@ -483,405 +970,18 @@ void cischeat_state::f1gpstr2_map(address_map &map)
 	// 0x100 RAM banks instead of 0x200
 	map(0x088000, 0x0887ff).ram();                                                                     // Linking with other units
 
-	map(0x090000, 0x097fff).ram().share("share2"); // Sharedram with sub CPU#2
-	map(0x098000, 0x09ffff).ram().share("share1"); // Sharedram with sub CPU#1
+	map(0x090000, 0x097fff).ram().share("road2_shareram"); // Sharedram with sub CPU#2
+	map(0x098000, 0x09ffff).ram().share("road1_shareram"); // Sharedram with sub CPU#1
 
 	/* Only writes to the first 0x40000 bytes affect the tilemaps:             */
 	/* either these games support larger tilemaps or have more ram than needed */
-	map(0x0a0000, 0x0a7fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0");     // Scroll ram 0
-	map(0x0a8000, 0x0affff).ram().w("scroll1", FUNC(megasys1_tilemap_device::write)).share("scroll1");     // Scroll ram 1
-	map(0x0b0000, 0x0b7fff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2");     // Scroll ram 2
+	map(0x0a0000, 0x0a7fff).ram().w(m_tmap[0], FUNC(megasys1_tilemap_device::write)).share("scroll0");     // Scroll ram 0
+	map(0x0a8000, 0x0affff).ram().w(m_tmap[1], FUNC(megasys1_tilemap_device::write)).share("scroll1");     // Scroll ram 1
+	map(0x0b0000, 0x0b7fff).ram().w(m_tmap[2], FUNC(megasys1_tilemap_device::write)).share("scroll2");     // Scroll ram 2
 	map(0x0b8000, 0x0bffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");               // Palettes
 
-	map(0x0f0000, 0x0fffff).ram().share("ram");                                             // RAM
+	map(0x0f0000, 0x0fffff).ram().share(m_ram);                                             // RAM
 	map(0x100000, 0x17ffff).rom().region("user1", 0);                                                            // ROM
-}
-
-
-/**************************************************************************
-                            Scud Hammer
-**************************************************************************/
-
-
-
-/*  Motor Status.
-
-    f--- ---- ---- ----     Rotation Limit (R?)
-    -e-- ---- ---- ----     Rotation Limit (L?)
-    --dc ba98 7654 32--
-    ---- ---- ---- --1-     Up Limit
-    ---- ---- ---- ---0     Down Limit  */
-
-uint16_t cischeat_state::scudhamm_motor_status_r()
-{
-	return m_scudhamm_motor_command;    // Motor Status
-}
-
-
-uint16_t cischeat_state::scudhamm_motor_pos_r()
-{
-	return 0x00 << 8;
-}
-
-
-/*  Move the motor.
-
-    fedc ba98 7654 32--
-    ---- ---- ---- --1-     Move Up
-    ---- ---- ---- ---0     Move Down
-
-    Within $20 vblanks the motor must reach the target. */
-
-void cischeat_state::scudhamm_motor_command_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	COMBINE_DATA( &m_scudhamm_motor_command );
-}
-
-
-uint8_t cischeat_state::scudhamm_analog_r()
-{
-	int i=ioport("IN1")->read(),j;
-
-	if ((i^m_prev)&0x4000) {
-		if (i<m_prev) m_prev-=0x8000;
-		else m_prev+=0x8000;
-	}
-
-	j=i-m_prev;
-	m_prev=i;
-
-	/* effect of hammer collision 'accelerometer':
-	$00 - $09 - no hit
-	$0A - $3F - soft hit
-	$40 - $FF - hard hit */
-	if (j<0) return 0;
-	else if (j>0xff) return 0xff;
-	return j;
-}
-
-/*
-    I don't know how many leds are there, but each bit in the buttons input
-    port (coins, tilt, buttons, select etc.) triggers the corresponding bit
-    in this word. I mapped the 3 buttons to the first 3 led.
-*/
-void cischeat_state::scudhamm_leds_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	if (ACCESSING_BITS_8_15)
-	{
-		m_leds[0] = BIT(data, 8);    // 3 buttons
-		m_leds[1] = BIT(data, 9);
-		m_leds[2] = BIT(data, 10);
-	}
-
-	if (ACCESSING_BITS_0_7)
-	{
-		machine().bookkeeping().coin_counter_w(0, data & 0x0001);
-		m_leds[3] = BIT(data, 4);
-		m_leds[4] = BIT(data, 5);
-	}
-}
-
-
-/*
-    $FFFC during self test, $FFFF onwards.
-    It could be audio(L/R) or layers(0/2) enable.
-*/
-void cischeat_state::scudhamm_enable_w(uint16_t data)
-{
-}
-
-
-void cischeat_state::scudhamm_oki_bank_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	if (ACCESSING_BITS_0_7)
-	{
-		m_oki1->set_rom_bank((data >> 0) & 0x3);
-		m_oki2->set_rom_bank((data >> 4) & 0x3);
-	}
-}
-
-void cischeat_state::scudhamm_map(address_map &map)
-{
-	map(0x000000, 0x07ffff).rom();                                                                 // ROM
-	map(0x082000, 0x082005).w("scroll0", FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).nopw(); //      UNUSED LAYER
-	map(0x082100, 0x082105).w("scroll2", FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082208, 0x082209).rw(m_watchdog, FUNC(watchdog_timer_device::reset16_r), FUNC(watchdog_timer_device::reset16_w));
-	map(0x0a0000, 0x0a3fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0");   // Scroll RAM 0
-	map(0x0b0000, 0x0b3fff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2");   // Scroll RAM 2
-	map(0x0b8000, 0x0bffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");          // Palette
-	map(0x0f0000, 0x0fffff).ram().share("ram");                                         // Work RAM + Spriteram
-	map(0x100000, 0x100001).w(FUNC(cischeat_state::scudhamm_oki_bank_w));                                          // Sound
-	map(0x100008, 0x100009).portr("IN0").w(FUNC(cischeat_state::scudhamm_leds_w));                          // Buttons
-	map(0x100015, 0x100015).rw(m_oki1, FUNC(okim6295_device::read), FUNC(okim6295_device::write));             // Sound
-	map(0x100019, 0x100019).rw(m_oki2, FUNC(okim6295_device::read), FUNC(okim6295_device::write));             //
-	map(0x10001c, 0x10001d).w(FUNC(cischeat_state::scudhamm_enable_w));                                            // ?
-	map(0x100041, 0x100041).rw("adc", FUNC(adc0804_device::read), FUNC(adc0804_device::write));                    // A / D
-	map(0x100044, 0x100045).r(FUNC(cischeat_state::scudhamm_motor_pos_r));                                  // Motor Position
-	map(0x100050, 0x100051).r(FUNC(cischeat_state::scudhamm_motor_status_r)).w(FUNC(cischeat_state::scudhamm_motor_command_w));        // Motor Limit Switches
-	map(0x10005c, 0x10005d).portr("IN2");                                                    // 2 x DSW
-}
-
-
-/**************************************************************************
-                            Arm Champs II
-
-The Arm Champs II start-up arm movement and info
----------------------------------------------------------------------
-At power-on, the arm moves left, then right then to center and stops.
-The center position is 0000.
-Note when the arm is at 0000 (center) the center limit switch is not
-on. It comes on when the arm is moved just off center (+-0002 or greater
-away from center) and stays on as long as the arm is not on
-0001-0000-FFFF (both left and right side). When the arm hits the
-left/right limit, both center and that left or right limit switch is on.
-
-Full movement to left limit switch is 0000-0040 (0x41h) and full right
-limit switch movement is 0000-FFC2 (0x3eh). Obviously the pot and arm
-are not 100% accurate and there's a small amount of slop between the
-arm shaft / motor mechanism and the pot.
-The limit switches are triggered just before the full movement.
-For the purpose of MAME emulation it can be rounded off so both sides
-move +- 0x40h.
-**************************************************************************/
-
-uint16_t armchamp2_state::motor_status_r()
-{
-	return 0x11;
-}
-
-void armchamp2_state::motor_command_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	// 0x00xx -> disable motor (test mode)
-	// 0x10ff -> automated test (test limits?)
-	// 0x10xx -> apply a left-to-right force
-	COMBINE_DATA( &m_arm_motor_command );
-}
-
-uint8_t armchamp2_state::analog_r()
-{
-	int armdelta;
-
-	armdelta = ioport("ARM")->read() - m_armold;
-	m_armold = ioport("ARM")->read();
-
-	return (~( m_arm_motor_command + armdelta )) & 0xff;    // + x : x<=0 and player loses, x>0 and player wins
-}
-
-/*
-    f--- ---- ---- ----
-    -edc ---- ---- ----     Leds
-    ---- ba9- ---- ----
-    ---- ---8 ---- ----     Start button led
-    ---- ---- 76-- ----     Coin counters
-    ---- ---- --54 3210
-*/
-void armchamp2_state::output_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	if (ACCESSING_BITS_8_15)
-	{
-		m_leds[0] = BIT(data, 8);
-		m_leds[1] = BIT(data, 12);
-		m_leds[2] = BIT(data, 13);
-		m_leds[3] = BIT(data, 14);
-	}
-
-	if (ACCESSING_BITS_0_7)
-	{
-		machine().bookkeeping().coin_counter_w(0, data & 0x0040);
-		machine().bookkeeping().coin_counter_w(1, data & 0x0080);
-	}
-}
-
-void armchamp2_state::armchmp2_map(address_map &map)
-{
-	map(0x000000, 0x07ffff).rom();
-	map(0x082000, 0x082005).w("scroll0", FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).nopw();
-	map(0x082100, 0x082105).w("scroll2", FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082208, 0x082209).rw(m_watchdog, FUNC(watchdog_timer_device::reset16_r), FUNC(watchdog_timer_device::reset16_w));
-	map(0x0a0000, 0x0a7fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0");
-	map(0x0b0000, 0x0b7fff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2");
-	map(0x0b8000, 0x0bffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");
-	map(0x0f0000, 0x0fffff).ram().share("ram"); // Work RAM + Spriteram
-	map(0x100000, 0x100001).portr("IN2").w(FUNC(armchamp2_state::scudhamm_oki_bank_w));
-	map(0x100004, 0x100005).portr("DSW"); // DSW
-	map(0x100008, 0x100009).portr("IN0").w(FUNC(armchamp2_state::output_w));
-	map(0x10000d, 0x10000d).rw("adc", FUNC(adc0804_device::read), FUNC(adc0804_device::write));
-	map(0x100010, 0x100011).rw(FUNC(armchamp2_state::motor_status_r), FUNC(armchamp2_state::motor_command_w));
-	// same hookup as acommand, most notably the "Arm Champs II" voice sample on title screen playbacks on oki1 mirror
-	map(0x100014, 0x100017).rw(m_oki1, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
-	map(0x100018, 0x10001b).rw(m_oki2, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
-}
-
-
-/**************************************************************************
-                                Captain Flag
-**************************************************************************/
-
-#define RIGHT 0
-#define LEFT  1
-
-void captflag_state::machine_start()
-{
-	cischeat_state::machine_start();
-
-	m_motor_left_output.resolve();
-	m_motor_right_output.resolve();
-}
-
-void captflag_state::leds_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	COMBINE_DATA( &m_captflag_leds );
-	if (ACCESSING_BITS_8_15)
-	{
-		machine().bookkeeping().coin_counter_w(1, data & 0x0100);    // coin 2
-		m_leds[0] = BIT(data, 8);    // decide
-		machine().bookkeeping().coin_counter_w(0, data & 0x0400);    // coin 1
-		m_leds[1] = BIT(data, 13);    // select
-
-		int power = (data & 0x1000);
-		m_hopper->motor_w(power ? 1 : 0);    // prize motor
-		if (!power)
-			m_hopper->reset();
-	}
-}
-
-void captflag_state::oki_bank_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	if (ACCESSING_BITS_0_7)
-	{
-		m_oki1_bank->set_entry(((data >> 0) + 1) & 0x7);
-		m_oki2_bank->set_entry(((data >> 4) + 1) & 0x7);
-	}
-}
-
-// Motors
-
-void captflag_state::motor_command_right_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	// Output check:
-	// e09a up
-	// 80b9 - (when not busy)
-	// 0088 - (when busy)
-	// e0ba down
-	data = COMBINE_DATA( &m_motor_command[RIGHT] );
-	motor_move(RIGHT, data);
-}
-void captflag_state::motor_command_left_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	// Output check:
-	// e0ba up
-	// 8099 - (when not busy)
-	// 0088 - (when busy)
-	// e09a down
-	data = COMBINE_DATA( &m_motor_command[LEFT] );
-	motor_move(LEFT, data);
-}
-
-void captflag_state::motor_move(int side, uint16_t data)
-{
-	uint16_t & pos  = m_motor_pos[side];
-
-	timer_device &dev((side == RIGHT) ? *m_motor_right : *m_motor_left);
-
-//  bool busy = !(dev.remaining() == attotime::never);
-	bool busy = false;
-
-	if (data & 0x0010)
-	{
-		if (!busy)
-		{
-			bool down = (data & 0x0020);
-
-			int inc;
-			switch (data >> 8)
-			{
-				case 0xf5:  inc = +2;   break;  // -5 -6
-				case 0xf8:  inc = +1;   break;  // -5 -3
-				case 0xfe:  inc = -1;   break;  // -5 +3
-				case 0x01:  inc = -2;   break;  // -5 +6
-				default:
-					if ((data >> 8) + 5 >= 0x100)
-						inc = -1;
-					else
-						inc = +1;
-			}
-
-			if (data & 0x0001)
-				inc = 1;
-
-			if (down)
-				inc *= -1;
-
-			int new_pos = pos + inc;
-			if (new_pos > 3)
-				new_pos = 3;
-			else if (new_pos < 0)
-				new_pos = 0;
-			busy = (new_pos != pos);
-			pos = new_pos;
-
-			if (busy)
-				dev.adjust(attotime::from_msec(100));
-		}
-	}
-	else
-	{
-		dev.reset();
-	}
-
-	((side == RIGHT) ? m_motor_right_output : m_motor_left_output) = pos;
-}
-
-template <int N>
-ioport_value captflag_state::motor_pos_r()
-{
-	const uint8_t pos[4] = {1,0,2,3}; // -> 2,3,1,0 offsets -> 0123
-	return ~pos[m_motor_pos[N]];
-}
-
-template <int N>
-int captflag_state::motor_busy_r()
-{
-//  timer_device & dev = ((side == RIGHT) ? m_motor_right : m_motor_left);
-//  return (dev.remaining() == attotime::never) ? 0 : 1;
-	return 0;
-}
-
-void captflag_state::captflag_map(address_map &map)
-{
-	map(0x000000, 0x07ffff).rom();                                                                 // ROM
-	map(0x082000, 0x082005).w("scroll0", FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082008, 0x08200d).nopw(); //      UNUSED LAYER
-	map(0x082100, 0x082105).w("scroll2", FUNC(megasys1_tilemap_device::scroll_w));
-	map(0x082208, 0x082209).rw(m_watchdog, FUNC(watchdog_timer_device::reset16_r), FUNC(watchdog_timer_device::reset16_w));
-	map(0x090008, 0x090009).nopw();                                                            // 0?
-	map(0x0a0000, 0x0a7fff).ram().w("scroll0", FUNC(megasys1_tilemap_device::write)).share("scroll0"); // Scroll RAM 0
-	map(0x0b0000, 0x0b7fff).ram().w("scroll2", FUNC(megasys1_tilemap_device::write)).share("scroll2"); // Scroll RAM 2
-	map(0x0b8000, 0x0bffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");  // Palette
-	map(0x0f0000, 0x0fffff).ram().share("ram");                                                 // Work RAM + Spriteram
-	map(0x100000, 0x100001).portr("SW1_2").w(FUNC(captflag_state::oki_bank_w));                    // 2 x DSW + Sound
-	map(0x100008, 0x100009).portr("Buttons").w(FUNC(captflag_state::leds_w));                      // Buttons + Leds
-	map(0x100015, 0x100015).rw(m_oki1, FUNC(okim6295_device::read), FUNC(okim6295_device::write));         // Sound
-	map(0x100019, 0x100019).rw(m_oki2, FUNC(okim6295_device::read), FUNC(okim6295_device::write));         //
-	map(0x10001c, 0x10001d).w(FUNC(captflag_state::scudhamm_enable_w));                                            // ?
-	map(0x100040, 0x100041).portr("SW01");                                                   // DSW + Motor
-	map(0x100044, 0x100045).w(FUNC(captflag_state::motor_command_left_w));                                 // Motor Command (Left)
-	map(0x100048, 0x100049).w(FUNC(captflag_state::motor_command_right_w));                                // Motor Command (Right)
-	map(0x100060, 0x10007f).ram().share("nvram");      // NVRAM (even bytes only)
-}
-
-void captflag_state::oki1_map(address_map &map)
-{
-	map(0x00000, 0x1ffff).rom();
-	map(0x20000, 0x3ffff).bankr("oki1_bank");
-}
-
-void captflag_state::oki2_map(address_map &map)
-{
-	map(0x00000, 0x1ffff).rom();
-	map(0x20000, 0x3ffff).bankr("oki2_bank");
 }
 
 
@@ -900,16 +1000,16 @@ void captflag_state::oki2_map(address_map &map)
 void cischeat_state::bigrun_map2(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                          // ROM
-	map(0x040000, 0x043fff).ram().share("share1");          // Shared RAM (with Main CPU)
-	map(0x080000, 0x0807ff).ram().share("roadram.0");       // Road RAM
+	map(0x040000, 0x043fff).ram().share("road1_shareram");  // Shared RAM (with Main CPU)
+	map(0x080000, 0x0807ff).ram().share(m_roadram[0]);      // Road RAM
 	map(0x0c0000, 0x0c3fff).ram();                          // RAM
 }
 
 void cischeat_state::bigrun_map3(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                          // ROM
-	map(0x040000, 0x043fff).ram().share("share2");          // Shared RAM (with Main CPU)
-	map(0x080000, 0x0807ff).ram().share("roadram.1");       // Road RAM
+	map(0x040000, 0x043fff).ram().share("road2_shareram");  // Shared RAM (with Main CPU)
+	map(0x080000, 0x0807ff).ram().share(m_roadram[1]);      // Road RAM
 	map(0x0c0000, 0x0c3fff).ram();                          // RAM
 }
 
@@ -921,21 +1021,21 @@ void cischeat_state::bigrun_map3(address_map &map)
 void cischeat_state::cischeat_map2(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                          // ROM
-	map(0x040000, 0x047fff).ram().share("share1");          // Shared RAM (with Main CPU)
-	map(0x080000, 0x0807ff).ram().share("roadram.0");       // Road RAM
+	map(0x040000, 0x047fff).ram().share("road1_shareram");  // Shared RAM (with Main CPU)
+	map(0x080000, 0x0807ff).ram().share(m_roadram[0]);      // Road RAM
 	map(0x0c0000, 0x0c3fff).ram();                          // RAM
 	map(0x100000, 0x100001).nopw();                         // watchdog
-	map(0x200000, 0x23ffff).rom().region("cpu2", 0x40000);  // ROM
+	map(0x200000, 0x23ffff).rom().region("road1cpu", 0x40000);  // ROM
 }
 
 void cischeat_state::cischeat_map3(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                          // ROM
-	map(0x040000, 0x047fff).ram().share("share2");          // Shared RAM (with Main CPU)
-	map(0x080000, 0x0807ff).ram().share("roadram.1");       // Road RAM
+	map(0x040000, 0x047fff).ram().share("road2_shareram");  // Shared RAM (with Main CPU)
+	map(0x080000, 0x0807ff).ram().share(m_roadram[1]);      // Road RAM
 	map(0x0c0000, 0x0c3fff).ram();                          // RAM
 	map(0x100000, 0x100001).nopw();                         // watchdog
-	map(0x200000, 0x23ffff).rom().region("cpu3", 0x40000);  // ROM
+	map(0x200000, 0x23ffff).rom().region("road2cpu", 0x40000);  // ROM
 }
 
 
@@ -946,18 +1046,18 @@ void cischeat_state::cischeat_map3(address_map &map)
 
 void cischeat_state::f1gpstar_map2(address_map &map)
 {
-	map(0x000000, 0x03ffff).rom();                          // ROM
-	map(0x080000, 0x087fff).ram().share("share1");          // Shared RAM (with Main CPU)
-	map(0x100000, 0x1007ff).ram().share("roadram.0");   // Road RAM
-	map(0x180000, 0x183fff).ram();                          // RAM
-	map(0x200000, 0x200001).nopw();                         // watchdog
+	map(0x000000, 0x03ffff).rom();                         // ROM
+	map(0x080000, 0x087fff).ram().share("road1_shareram"); // Shared RAM (with Main CPU)
+	map(0x100000, 0x1007ff).ram().share(m_roadram[0]);     // Road RAM
+	map(0x180000, 0x183fff).ram();                         // RAM
+	map(0x200000, 0x200001).nopw();                        // watchdog
 }
 
 void cischeat_state::f1gpstar_map3(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                          // ROM
-	map(0x080000, 0x087fff).ram().share("share2");          // Shared RAM (with Main CPU)
-	map(0x100000, 0x1007ff).ram().share("roadram.1");       // Road RAM
+	map(0x080000, 0x087fff).ram().share("road2_shareram");  // Shared RAM (with Main CPU)
+	map(0x100000, 0x1007ff).ram().share(m_roadram[1]);      // Road RAM
 	map(0x180000, 0x183fff).ram();                          // RAM
 	map(0x200000, 0x200001).nopw();                         // watchdog
 }
@@ -978,23 +1078,23 @@ void cischeat_state::f1gpstar_map3(address_map &map)
                                 Big Run
 **************************************************************************/
 
-void cischeat_state::bigrun_soundbank_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+void cischeat_state::bigrun_soundbank_w(offs_t offset, u16 data, u16 mem_mask)
 {
 	if (ACCESSING_BITS_0_7)
 	{
-		m_oki1->set_rom_bank((data >> 0) & 1);
-		m_oki2->set_rom_bank((data >> 4) & 1);
+		m_oki[0]->set_rom_bank(BIT(data, 0));
+		m_oki[1]->set_rom_bank(BIT(data, 4));
 	}
 }
 
 void cischeat_state::bigrun_sound_map(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                                                 // ROM
-	map(0x040000, 0x040001).r(m_soundlatch, FUNC(generic_latch_16_device::read)).w(FUNC(cischeat_state::bigrun_soundbank_w));    // From Main CPU
-	map(0x060000, 0x060001).w(m_soundlatch2, FUNC(generic_latch_16_device::write));                           // To Main CPU
+	map(0x040000, 0x040001).r(m_soundlatch[0], FUNC(generic_latch_16_device::read)).w(FUNC(cischeat_state::bigrun_soundbank_w));    // From Main CPU
+	map(0x060000, 0x060001).w(m_soundlatch[1], FUNC(generic_latch_16_device::write));                           // To Main CPU
 	map(0x080000, 0x080003).rw("ymsnd", FUNC(ym2151_device::read), FUNC(ym2151_device::write)).umask16(0x00ff);
-	map(0x0a0000, 0x0a0003).rw(m_oki1, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
-	map(0x0c0000, 0x0c0003).rw(m_oki2, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0a0000, 0x0a0003).rw(m_oki[0], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0c0000, 0x0c0003).rw(m_oki[1], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
 	map(0x0f0000, 0x0fffff).ram();                                                 // RAM
 }
 
@@ -1003,26 +1103,23 @@ void cischeat_state::bigrun_sound_map(address_map &map)
                                 Cisco Heat
 **************************************************************************/
 
-void cischeat_state::cischeat_soundbank_1_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+template <unsigned Which>
+void cischeat_state::cischeat_soundbank_w(offs_t offset, u16 data, u16 mem_mask)
 {
-	if (ACCESSING_BITS_0_7) m_oki1->set_rom_bank(data & 1);
-}
-
-void cischeat_state::cischeat_soundbank_2_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	if (ACCESSING_BITS_0_7) m_oki2->set_rom_bank(data & 1);
+	if (ACCESSING_BITS_0_7)
+		m_oki[Which]->set_rom_bank(BIT(data, 0));
 }
 
 void cischeat_state::cischeat_sound_map(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                                                 // ROM
-	map(0x040002, 0x040003).w(FUNC(cischeat_state::cischeat_soundbank_1_w));               // Sample Banking
-	map(0x040004, 0x040005).w(FUNC(cischeat_state::cischeat_soundbank_2_w));               // Sample Banking
-	map(0x060002, 0x060003).w(m_soundlatch2, FUNC(generic_latch_16_device::write));                          // To Main CPU
-	map(0x060004, 0x060005).r(m_soundlatch, FUNC(generic_latch_16_device::read));                             // From Main CPU
+	map(0x040002, 0x040003).w(FUNC(cischeat_state::cischeat_soundbank_w<0>));               // Sample Banking
+	map(0x040004, 0x040005).w(FUNC(cischeat_state::cischeat_soundbank_w<1>));               // Sample Banking
+	map(0x060002, 0x060003).w(m_soundlatch[1], FUNC(generic_latch_16_device::write));                          // To Main CPU
+	map(0x060004, 0x060005).r(m_soundlatch[0], FUNC(generic_latch_16_device::read));                             // From Main CPU
 	map(0x080000, 0x080003).rw("ymsnd", FUNC(ym2151_device::read), FUNC(ym2151_device::write)).umask16(0x00ff);
-	map(0x0a0000, 0x0a0003).rw(m_oki1, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
-	map(0x0c0000, 0x0c0003).rw(m_oki2, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0a0000, 0x0a0003).rw(m_oki[0], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0c0000, 0x0c0003).rw(m_oki[1], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
 	map(0x0f0000, 0x0fffff).ram();                                                 // RAM
 }
 
@@ -1034,12 +1131,12 @@ void cischeat_state::cischeat_sound_map(address_map &map)
 void cischeat_state::f1gpstar_sound_map(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom();                                                 // ROM
-	map(0x040004, 0x040005).w(FUNC(cischeat_state::cischeat_soundbank_1_w));               // Sample Banking   (cischeat: 40002)
-	map(0x040008, 0x040009).w(FUNC(cischeat_state::cischeat_soundbank_2_w));               // Sample Banking   (cischeat: 40004)
-	map(0x060000, 0x060001).r(m_soundlatch, FUNC(generic_latch_16_device::read)).w(m_soundlatch2, FUNC(generic_latch_16_device::write));   // From Main CPU    (cischeat: 60004)
+	map(0x040004, 0x040005).w(FUNC(cischeat_state::cischeat_soundbank_w<0>));               // Sample Banking   (cischeat: 40002)
+	map(0x040008, 0x040009).w(FUNC(cischeat_state::cischeat_soundbank_w<1>));               // Sample Banking   (cischeat: 40004)
+	map(0x060000, 0x060001).r(m_soundlatch[0], FUNC(generic_latch_16_device::read)).w(m_soundlatch[1], FUNC(generic_latch_16_device::write));   // From Main CPU    (cischeat: 60004)
 	map(0x080000, 0x080003).rw("ymsnd", FUNC(ym2151_device::read), FUNC(ym2151_device::write)).umask16(0x00ff);
-	map(0x0a0000, 0x0a0003).rw(m_oki1, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
-	map(0x0c0000, 0x0c0003).rw(m_oki2, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0a0000, 0x0a0003).rw(m_oki[0], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0c0000, 0x0c0003).rw(m_oki[1], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
 	map(0x0e0000, 0x0fffff).ram();                                                 // RAM              (cischeat: f0000-fffff)
 }
 
@@ -1051,14 +1148,14 @@ void cischeat_state::f1gpstar_sound_map(address_map &map)
 void cischeat_state::f1gpstr2_sound_map(address_map &map)
 {
 	map(0x000000, 0x03ffff).rom(); // ROM
-	map(0x040004, 0x040005).w(FUNC(cischeat_state::cischeat_soundbank_1_w));                   // Sample Banking
-	map(0x040008, 0x040009).w(FUNC(cischeat_state::cischeat_soundbank_2_w));                   // Sample Banking
+	map(0x040004, 0x040005).w(FUNC(cischeat_state::cischeat_soundbank_w<0>));                   // Sample Banking
+	map(0x040008, 0x040009).w(FUNC(cischeat_state::cischeat_soundbank_w<1>));                   // Sample Banking
 	map(0x04000e, 0x04000f).nopw();                                            // ? 0              (f1gpstar: no)
-	map(0x060002, 0x060003).w(m_soundlatch2, FUNC(generic_latch_16_device::write));                          // To Main CPU
-	map(0x060004, 0x060005).r(m_soundlatch, FUNC(generic_latch_16_device::read));                             // From Main CPU
+	map(0x060002, 0x060003).w(m_soundlatch[1], FUNC(generic_latch_16_device::write));                          // To Main CPU
+	map(0x060004, 0x060005).r(m_soundlatch[0], FUNC(generic_latch_16_device::read));                             // From Main CPU
 	map(0x080000, 0x080003).rw("ymsnd", FUNC(ym2151_device::read), FUNC(ym2151_device::write)).umask16(0x00ff);
-	map(0x0a0000, 0x0a0003).rw(m_oki1, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
-	map(0x0c0000, 0x0c0003).rw(m_oki2, FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0a0000, 0x0a0003).rw(m_oki[0], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
+	map(0x0c0000, 0x0c0003).rw(m_oki[1], FUNC(okim6295_device::read), FUNC(okim6295_device::write)).umask16(0x00ff);
 	map(0x0e0000, 0x0fffff).ram();                                                     // RAM
 }
 
@@ -1072,10 +1169,10 @@ void cischeat_state::f1gpstr2_sound_map(address_map &map)
 
 void cischeat_state::f1gpstr2_io_map(address_map &map)
 {
-	map(0x000000, 0x07ffff).rom();                                         // ROM
-	map(0x080000, 0x080fff).ram().share("shareio");
-	map(0x100000, 0x100001).writeonly().share("ioready");   //
-	map(0x180000, 0x183fff).ram();                                         // RAM
+	map(0x000000, 0x07ffff).rom();                                  // ROM
+	map(0x080000, 0x080fff).ram().share("io_shareram");
+	map(0x100000, 0x100001).writeonly().share(m_f1gpstr2_ioready);  //
+	map(0x180000, 0x183fff).ram();                                  // RAM
 	map(0x200000, 0x200001).noprw();                                //
 }
 
@@ -1592,363 +1689,6 @@ INPUT_PORTS_END
 
 
 /**************************************************************************
-                                Scud Hammer
-**************************************************************************/
-
-static INPUT_PORTS_START( scudhamma )
-	PORT_START("IN0")   // Buttons
-	PORT_BIT( 0x0001, IP_ACTIVE_LOW, IPT_COIN1 ) // PORT_IMPULSE(2)
-	PORT_BIT( 0x0002, IP_ACTIVE_LOW, IPT_UNKNOWN  ) // GAME OVER if pressed on the selection screen
-	PORT_BIT( 0x0004, IP_ACTIVE_LOW, IPT_SERVICE1 )
-	PORT_SERVICE_NO_TOGGLE( 0x0008, IP_ACTIVE_LOW )     // called "Test"
-	PORT_BIT( 0x0010, IP_ACTIVE_LOW, IPT_START1 )
-	PORT_BIT( 0x0020, IP_ACTIVE_LOW, IPT_BUTTON4 ) // Select
-	PORT_BIT( 0x0040, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0080, IP_ACTIVE_LOW, IPT_UNKNOWN )
-
-	PORT_BIT( 0x0100, IP_ACTIVE_HIGH, IPT_BUTTON1 ) // Gu (rock)
-	PORT_BIT( 0x0200, IP_ACTIVE_HIGH, IPT_BUTTON2 ) // Choki (scissors)
-	PORT_BIT( 0x0400, IP_ACTIVE_HIGH, IPT_BUTTON3 ) // Pa (paper)
-	PORT_BIT( 0x0800, IP_ACTIVE_LOW,  IPT_UNKNOWN )
-	PORT_BIT( 0x1000, IP_ACTIVE_LOW,  IPT_UNKNOWN )
-	PORT_BIT( 0x2000, IP_ACTIVE_LOW,  IPT_UNKNOWN )
-	PORT_BIT( 0x4000, IP_ACTIVE_HIGH, IPT_TILT )
-	PORT_BIT( 0x8000, IP_ACTIVE_LOW,  IPT_UNKNOWN )
-
-	PORT_START("IN1")   // A/D
-	PORT_BIT( 0x7fff, 0x0000, IPT_DIAL_V ) PORT_SENSITIVITY(50) PORT_KEYDELTA(50)
-
-	PORT_START("IN2")   // DSW
-	PORT_DIPNAME( 0x0003, 0x0003, DEF_STR( Difficulty ) ) PORT_DIPLOCATION("SW3:1,2")
-	PORT_DIPSETTING(      0x0000, DEF_STR( Easy ) )
-	PORT_DIPSETTING(      0x0003, DEF_STR( Normal ) )
-	PORT_DIPSETTING(      0x0002, DEF_STR( Hard ) )
-	PORT_DIPSETTING(      0x0001, DEF_STR( Hardest ) )
-	PORT_DIPNAME( 0x000c, 0x000c, "Time To Hit" ) PORT_DIPLOCATION("SW3:3,4")
-	PORT_DIPSETTING(      0x000c, "2 s" )
-	PORT_DIPSETTING(      0x0008, "3 s" )
-	PORT_DIPSETTING(      0x0004, "4 s" )
-	PORT_DIPSETTING(      0x0000, "5 s" )
-	PORT_DIPNAME( 0x0010, 0x0010, DEF_STR( Demo_Sounds ) ) PORT_DIPLOCATION("SW3:5")
-	PORT_DIPSETTING(      0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(      0x0010, DEF_STR( On ) )
-	PORT_DIPNAME( 0x0020, 0x0020, DEF_STR( Lives ) ) PORT_DIPLOCATION("SW3:6")
-	PORT_DIPSETTING(      0x0020, "3" )
-	PORT_DIPSETTING(      0x0000, "5" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0040, 0x0040, "SW3:7" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0080, 0x0080, "SW3:8" )
-
-	PORT_DIPNAME( 0x0700, 0x0700, DEF_STR( Coinage ) ) PORT_DIPLOCATION("SW4:1,2,3")
-	PORT_DIPSETTING(      0x0100, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(      0x0200, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(      0x0300, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(      0x0700, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(      0x0600, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(      0x0500, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(      0x0400, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( Free_Play ) )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0800, 0x0800, "SW4:4" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x1000, 0x1000, "SW4:5" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x2000, 0x2000, "SW4:6" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x4000, 0x4000, "SW4:7" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x8000, 0x8000, "SW4:8" )
-INPUT_PORTS_END
-
-static INPUT_PORTS_START( scudhamm )
-	PORT_INCLUDE( scudhamma )
-
-	PORT_MODIFY("IN2")
-	PORT_DIPNAME( 0x0800, 0x0800, "Win Up" ) PORT_DIPLOCATION("SW4:4")
-	PORT_DIPSETTING(      0x0800, DEF_STR( Off ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( On ) )
-INPUT_PORTS_END
-
-
-/**************************************************************************
-                            Arm Champs II
-**************************************************************************/
-
-ioport_value armchamp2_state::left_sensor_r()
-{
-	int arm_x = ioport("ARM")->read();
-	return (arm_x < 0x40);
-}
-
-ioport_value armchamp2_state::right_sensor_r()
-{
-	int arm_x = ioport("ARM")->read();
-	return (arm_x > 0xc0);
-}
-
-ioport_value armchamp2_state::center_sensor_r()
-{
-	int arm_x = ioport("ARM")->read();
-	return ((arm_x > 0x60) && (arm_x < 0xa0));
-}
-
-
-static INPUT_PORTS_START( armchmp2 )
-	PORT_START("IN0")   // Buttons + Sensors
-	PORT_BIT( 0x0001, IP_ACTIVE_LOW, IPT_CUSTOM ) PORT_CUSTOM_MEMBER(FUNC(armchamp2_state::left_sensor_r))
-	PORT_BIT( 0x0002, IP_ACTIVE_LOW, IPT_CUSTOM ) PORT_CUSTOM_MEMBER(FUNC(armchamp2_state::right_sensor_r))
-	PORT_BIT( 0x0004, IP_ACTIVE_LOW, IPT_CUSTOM ) PORT_CUSTOM_MEMBER(FUNC(armchamp2_state::center_sensor_r))
-	PORT_BIT( 0x0008, IP_ACTIVE_LOW, IPT_SERVICE1 )
-	PORT_SERVICE_NO_TOGGLE( 0x0010, IP_ACTIVE_LOW )
-	PORT_BIT( 0x0020, IP_ACTIVE_LOW, IPT_COIN1 )
-	PORT_BIT( 0x0040, IP_ACTIVE_LOW, IPT_COIN2 )
-	PORT_BIT( 0x0080, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0100, IP_ACTIVE_LOW, IPT_START1 )
-	PORT_BIT( 0x0200, IP_ACTIVE_LOW, IPT_BUTTON1 ) // hard
-	PORT_BIT( 0x0400, IP_ACTIVE_LOW, IPT_BUTTON2 ) // easy
-	PORT_BIT( 0x0800, IP_ACTIVE_LOW, IPT_BUTTON3 ) // elbow (it always complains though)
-
-	PORT_START("ARM")   // A/D
-	PORT_BIT( 0xff, 0x00, IPT_DIAL ) PORT_MINMAX(0x00,0xff) PORT_SENSITIVITY(100) PORT_KEYDELTA(10)
-
-	PORT_START("IN2")   // DSW
-	PORT_DIPNAME( 0x0003, 0x0003, DEF_STR( Difficulty ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( Easy ) )
-	PORT_DIPSETTING(      0x0003, DEF_STR( Normal ) )
-	PORT_DIPSETTING(      0x0002, DEF_STR( Hard ) )
-	PORT_DIPSETTING(      0x0001, DEF_STR( Hardest ) )
-	PORT_DIPNAME( 0x000c, 0x000c, DEF_STR( Lives ) )
-	PORT_DIPSETTING(      0x0000, "1" )
-	PORT_DIPSETTING(      0x000c, "2" )
-	PORT_DIPSETTING(      0x0008, "3" )
-	PORT_DIPSETTING(      0x0004, "4" )
-	PORT_DIPNAME( 0x0030, 0x0030, DEF_STR( Game_Time ) )
-	PORT_DIPSETTING(      0x0000, "15 s" )
-	PORT_DIPSETTING(      0x0030, "20 s" )
-	PORT_DIPSETTING(      0x0020, "25 s" )
-	PORT_DIPSETTING(      0x0010, "30 s" )
-	PORT_DIPNAME( 0x00c0, 0x0000, DEF_STR( Region ) )   // some text is in japanese regardless!
-	PORT_DIPSETTING(      0x00c0, DEF_STR( Japan ) )
-	PORT_DIPSETTING(      0x0080, DEF_STR( USA ) )
-	PORT_DIPSETTING(      0x0040, DEF_STR( Europe ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( World ) )
-
-	PORT_START("DSW")
-	PORT_DIPNAME( 0x07, 0x07, DEF_STR( Coin_A ) )
-	PORT_DIPSETTING(    0x01, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(    0x02, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(    0x03, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(    0x07, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(    0x06, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(    0x05, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(    0x04, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(    0x00, DEF_STR( Free_Play ) )
-	PORT_DIPNAME( 0x38, 0x38, DEF_STR( Coin_B ) )
-	PORT_DIPSETTING(    0x08, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(    0x10, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(    0x18, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(    0x38, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(    0x30, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(    0x28, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(    0x20, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(    0x00, DEF_STR( Free_Play ) )
-	PORT_DIPNAME( 0x40, 0x40, DEF_STR( Demo_Sounds ) )
-	PORT_DIPSETTING(    0x00, DEF_STR( Off ) )
-	PORT_DIPSETTING(    0x40, DEF_STR( On ) )
-	PORT_DIPNAME( 0x80, 0x80, "Demo Arm" )
-	PORT_DIPSETTING(    0x80, "Center" )
-	PORT_DIPSETTING(    0x00, "Side" )
-INPUT_PORTS_END
-
-
-/**************************************************************************
-                                Captain Flag
-**************************************************************************/
-
-static INPUT_PORTS_START( captflag )
-	PORT_START("Buttons")
-	PORT_BIT( 0x0001, IP_ACTIVE_LOW, IPT_JOYSTICKLEFT_UP   ) PORT_NAME("Left Red Stick Up")
-	PORT_BIT( 0x0002, IP_ACTIVE_LOW, IPT_JOYSTICKLEFT_DOWN ) PORT_NAME("Left Red Stick Down")
-	PORT_BIT( 0x0004, IP_ACTIVE_LOW, IPT_JOYSTICKRIGHT_UP   ) PORT_NAME("Right White Stick Up")
-	PORT_BIT( 0x0008, IP_ACTIVE_LOW, IPT_JOYSTICKRIGHT_DOWN ) PORT_NAME("Right White Stick Down")
-	PORT_BIT( 0x0010, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0020, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0040, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0080, IP_ACTIVE_LOW, IPT_UNKNOWN )
-
-	PORT_BIT( 0x0100, IP_ACTIVE_LOW, IPT_COIN2    ) PORT_IMPULSE(1) // coin 2
-	PORT_BIT( 0x0200, IP_ACTIVE_LOW, IPT_BUTTON2  ) PORT_NAME("Decide Button") // decide
-	PORT_BIT( 0x0400, IP_ACTIVE_LOW, IPT_COIN1    ) PORT_IMPULSE(1) // coin 1
-	PORT_BIT( 0x0800, IP_ACTIVE_LOW, IPT_SERVICE1 ) // service
-	PORT_BIT( 0x1000, IP_ACTIVE_LOW, IPT_SERVICE  ) // test
-	PORT_BIT( 0x2000, IP_ACTIVE_LOW, IPT_BUTTON1  ) PORT_NAME("Select Button") // select
-	PORT_BIT( 0x4000, IP_ACTIVE_HIGH,IPT_OUTPUT   ) PORT_READ_LINE_DEVICE_MEMBER("hopper", FUNC(ticket_dispenser_device::line_r)) // prize sensor
-	PORT_BIT( 0x8000, IP_ACTIVE_LOW, IPT_UNKNOWN  ) // potery on schems?
-
-	PORT_START("SW1_2")
-	PORT_DIPNAME( 0x0007, 0x0007, DEF_STR( Coin_A ) ) PORT_DIPLOCATION("SW1:1,2,3")
-	PORT_DIPSETTING(      0x0001, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(      0x0002, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(      0x0003, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(      0x0007, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(      0x0006, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(      0x0005, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(      0x0004, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( Free_Play ) )
-	PORT_DIPNAME( 0x0038, 0x0038, DEF_STR( Coin_B ) ) PORT_DIPLOCATION("SW1:4,5,6")
-	PORT_DIPSETTING(      0x0008, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(      0x0010, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(      0x0018, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(      0x0038, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(      0x0030, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(      0x0028, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(      0x0020, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( Free_Play ) )
-	PORT_DIPNAME( 0x00c0, 0x00c0, "Qualify" ) PORT_DIPLOCATION("SW1:7,8")
-	PORT_DIPSETTING(      0x0000, "30/40" )
-	PORT_DIPSETTING(      0x00c0, "30/50" )
-	PORT_DIPSETTING(      0x0080, "40/50" )
-	PORT_DIPSETTING(      0x0040, "50/50" )
-
-	PORT_DIPUNKNOWN_DIPLOC( 0x0100, 0x0100, "SW2:1" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0200, 0x0200, "SW2:2" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0400, 0x0400, "SW2:3" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0800, 0x0800, "SW2:4" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x1000, 0x1000, "SW2:5" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x2000, 0x2000, "SW2:6" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x4000, 0x4000, "SW2:7" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x8000, 0x8000, "SW2:8" )
-
-	PORT_START("SW01")
-	PORT_BIT( 0x0003, IP_ACTIVE_LOW, IPT_CUSTOM) PORT_CUSTOM_MEMBER(FUNC(captflag_state::motor_pos_r<LEFT>))
-	PORT_BIT( 0x000c, IP_ACTIVE_LOW, IPT_CUSTOM) PORT_CUSTOM_MEMBER(FUNC(captflag_state::motor_pos_r<RIGHT>))
-
-	PORT_BIT( 0x0040, IP_ACTIVE_HIGH, IPT_CUSTOM) PORT_READ_LINE_MEMBER(FUNC(captflag_state::motor_busy_r<LEFT>))
-	PORT_BIT( 0x0080, IP_ACTIVE_HIGH, IPT_CUSTOM) PORT_READ_LINE_MEMBER(FUNC(captflag_state::motor_busy_r<RIGHT>))
-
-	PORT_DIPNAME( 0x0300, 0x0300, DEF_STR( Difficulty ) ) PORT_DIPLOCATION("SW01:1,2")
-	PORT_DIPSETTING(      0x0000, DEF_STR( Easy ) )
-	PORT_DIPSETTING(      0x0300, DEF_STR( Normal ) )
-	PORT_DIPSETTING(      0x0200, DEF_STR( Hard ) )
-	PORT_DIPSETTING(      0x0100, DEF_STR( Hardest ) )
-	PORT_DIPNAME( 0x0400, 0x0400, DEF_STR( Lives ) ) PORT_DIPLOCATION("SW01:3")
-	PORT_DIPSETTING(      0x0400, "3/4" )
-	PORT_DIPSETTING(      0x0000, "4/5" )
-	PORT_DIPNAME( 0x0800, 0x0800, "BGM Volume" ) PORT_DIPLOCATION("SW01:4")
-	PORT_DIPSETTING(      0x0000, "Low" )
-	PORT_DIPSETTING(      0x0800, DEF_STR( Normal ) )
-	PORT_DIPNAME( 0x1000, 0x1000, DEF_STR( Demo_Sounds ) ) PORT_DIPLOCATION("SW01:5")
-	PORT_DIPSETTING(      0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(      0x1000, DEF_STR( On ) )
-	PORT_DIPNAME( 0xe000, 0xe000, "Capsule Payout" ) PORT_DIPLOCATION("SW01:6,7,8")
-	PORT_DIPSETTING(      0x0000, "Nothing" )
-	PORT_DIPSETTING(      0xc000, "05%" )
-	PORT_DIPSETTING(      0xe000, "10%" )
-	PORT_DIPSETTING(      0xa000, "15%" )
-	PORT_DIPSETTING(      0x8000, "20%" )
-	PORT_DIPSETTING(      0x6000, "30%" )
-	PORT_DIPSETTING(      0x4000, "50%" )
-	PORT_DIPSETTING(      0x2000, "All" )
-INPUT_PORTS_END
-
-static INPUT_PORTS_START( vscaptfl )
-	PORT_START("Buttons")
-	PORT_BIT( 0x0001, IP_ACTIVE_LOW, IPT_JOYSTICKLEFT_UP   ) PORT_NAME("P1 Left Red Stick Up") PORT_PLAYER(1)
-	PORT_BIT( 0x0002, IP_ACTIVE_LOW, IPT_JOYSTICKLEFT_DOWN ) PORT_NAME("P1 Left Red Stick Down") PORT_PLAYER(1)
-	PORT_BIT( 0x0004, IP_ACTIVE_LOW, IPT_JOYSTICKRIGHT_UP   ) PORT_NAME("P1 Right White Stick Up") PORT_PLAYER(1)
-	PORT_BIT( 0x0008, IP_ACTIVE_LOW, IPT_JOYSTICKRIGHT_DOWN ) PORT_NAME("P1 Right White Stick Down") PORT_PLAYER(1)
-	PORT_BIT( 0x0010, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0020, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0040, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x0080, IP_ACTIVE_LOW, IPT_UNKNOWN )
-
-	PORT_BIT( 0x0100, IP_ACTIVE_LOW, IPT_COIN2    ) PORT_IMPULSE(1) // coin 2
-	PORT_BIT( 0x0200, IP_ACTIVE_LOW, IPT_BUTTON2  ) PORT_NAME("Decide Button") // decide
-	PORT_BIT( 0x0400, IP_ACTIVE_LOW, IPT_COIN1    ) PORT_IMPULSE(1) // coin 1
-	PORT_BIT( 0x0800, IP_ACTIVE_LOW, IPT_SERVICE1 ) // service
-	PORT_BIT( 0x1000, IP_ACTIVE_LOW, IPT_SERVICE  ) // test
-	PORT_BIT( 0x2000, IP_ACTIVE_LOW, IPT_BUTTON1  ) PORT_NAME("Select Button") // select
-	PORT_BIT( 0x4000, IP_ACTIVE_HIGH,IPT_OUTPUT   ) PORT_READ_LINE_DEVICE_MEMBER("hopper", FUNC(ticket_dispenser_device::line_r)) // prize sensor
-	PORT_BIT( 0x8000, IP_ACTIVE_LOW, IPT_UNKNOWN  ) // potery on schems?
-
-	PORT_START("Buttons2")
-	PORT_BIT( 0x0003, IP_ACTIVE_LOW, IPT_CUSTOM) PORT_CUSTOM_MEMBER(FUNC(captflag_state::motor_pos_r<LEFT>))
-	PORT_BIT( 0x000c, IP_ACTIVE_LOW, IPT_CUSTOM) PORT_CUSTOM_MEMBER(FUNC(captflag_state::motor_pos_r<RIGHT>))
-	PORT_BIT( 0x0040, IP_ACTIVE_HIGH, IPT_CUSTOM) PORT_READ_LINE_MEMBER(FUNC(captflag_state::motor_busy_r<LEFT>))
-	PORT_BIT( 0x0080, IP_ACTIVE_HIGH, IPT_CUSTOM) PORT_READ_LINE_MEMBER(FUNC(captflag_state::motor_busy_r<RIGHT>))
-	PORT_BIT( 0x0100, IP_ACTIVE_LOW, IPT_JOYSTICKLEFT_UP ) PORT_NAME("P2 Left Red Stick Up") PORT_PLAYER(2)
-	PORT_BIT( 0x0200, IP_ACTIVE_LOW, IPT_JOYSTICKLEFT_DOWN ) PORT_NAME("P2 Left Red Stick Down") PORT_PLAYER(2)
-	PORT_BIT( 0x0400, IP_ACTIVE_LOW, IPT_JOYSTICKRIGHT_UP ) PORT_NAME("P2 Right White Stick Up") PORT_PLAYER(2)
-	PORT_BIT( 0x0800, IP_ACTIVE_LOW, IPT_JOYSTICKRIGHT_DOWN ) PORT_NAME("P2 Right White Stick Down") PORT_PLAYER(2)
-	PORT_BIT( 0x1000, IP_ACTIVE_LOW, IPT_START ) PORT_PLAYER(1)
-	PORT_BIT( 0x2000, IP_ACTIVE_LOW, IPT_START ) PORT_PLAYER(2)
-	PORT_BIT( 0x4000, IP_ACTIVE_LOW, IPT_UNKNOWN )
-	PORT_BIT( 0x8000, IP_ACTIVE_LOW, IPT_UNKNOWN )
-
-	PORT_START("SW01")
-	PORT_DIPNAME( 0x0003, 0x0003, DEF_STR( Difficulty ) ) PORT_DIPLOCATION("SW01:1,2")
-	PORT_DIPSETTING(      0x0000, DEF_STR( Easy ) )
-	PORT_DIPSETTING(      0x0003, DEF_STR( Normal ) )
-	PORT_DIPSETTING(      0x0002, DEF_STR( Hard ) )
-	PORT_DIPSETTING(      0x0001, DEF_STR( Hardest ) )
-	PORT_DIPNAME( 0x0004, 0x0004, "Mode") PORT_DIPLOCATION("SW01:3")
-	PORT_DIPSETTING(      0x0004, "1 set" )
-	PORT_DIPSETTING(      0x0000, "3 set" )
-	PORT_DIPNAME( 0x0008, 0x0008, "BGM Volume" ) PORT_DIPLOCATION("SW01:4")
-	PORT_DIPSETTING(      0x0000, "Low" )
-	PORT_DIPSETTING(      0x0008, DEF_STR( Normal ) )
-	PORT_DIPNAME( 0x0010, 0x0010, DEF_STR( Demo_Sounds ) ) PORT_DIPLOCATION("SW01:5")
-	PORT_DIPSETTING(      0x0000, DEF_STR( Off ) )
-	PORT_DIPSETTING(      0x0010, DEF_STR( On ) )
-	PORT_DIPNAME( 0x00e0, 0x00e0, "Capsule Payout" ) PORT_DIPLOCATION("SW01:6,7,8")
-	PORT_DIPSETTING(      0x0000, "Nothing" )
-	PORT_DIPSETTING(      0x00c0, "05%" )
-	PORT_DIPSETTING(      0x00e0, "10%" )
-	PORT_DIPSETTING(      0x00a0, "15%" )
-	PORT_DIPSETTING(      0x0080, "20%" )
-	PORT_DIPSETTING(      0x0060, "30%" )
-	PORT_DIPSETTING(      0x0040, "50%" )
-	PORT_DIPSETTING(      0x0020, "All" )
-
-	PORT_DIPNAME( 0x0700, 0x0700, DEF_STR( Coin_A ) ) PORT_DIPLOCATION("SW02:1,2,3")
-	PORT_DIPSETTING(      0x0100, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(      0x0200, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(      0x0300, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(      0x0700, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(      0x0600, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(      0x0500, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(      0x0400, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( Free_Play ) )
-	PORT_DIPNAME( 0x3800, 0x3800, DEF_STR( Coin_B ) ) PORT_DIPLOCATION("SW02:4,5,6")
-	PORT_DIPSETTING(      0x0800, DEF_STR( 4C_1C ) )
-	PORT_DIPSETTING(      0x1000, DEF_STR( 3C_1C ) )
-	PORT_DIPSETTING(      0x1800, DEF_STR( 2C_1C ) )
-	PORT_DIPSETTING(      0x3800, DEF_STR( 1C_1C ) )
-	PORT_DIPSETTING(      0x3000, DEF_STR( 1C_2C ) )
-	PORT_DIPSETTING(      0x2800, DEF_STR( 1C_3C ) )
-	PORT_DIPSETTING(      0x2000, DEF_STR( 1C_4C ) )
-	PORT_DIPSETTING(      0x0000, DEF_STR( Free_Play ) )
-	PORT_DIPNAME( 0xc000, 0xc000, "Qualify" ) PORT_DIPLOCATION("SW02:7,8")
-	PORT_DIPSETTING(      0x0000, "30/40" )
-	PORT_DIPSETTING(      0xc000, "30/50" )
-	PORT_DIPSETTING(      0x8000, "40/50" )
-	PORT_DIPSETTING(      0x4000, "50/50" )
-
-	PORT_START("SW1_2")
-	PORT_DIPUNKNOWN_DIPLOC( 0x0001, 0x0001, "SW1:1" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0002, 0x0002, "SW1:2" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0004, 0x0004, "SW1:3" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0008, 0x0008, "SW1:4" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0010, 0x0010, "SW1:5" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0020, 0x0020, "SW1:6" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0040, 0x0040, "SW1:7" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0080, 0x0080, "SW1:8" )
-
-	PORT_DIPUNKNOWN_DIPLOC( 0x0100, 0x0100, "SW2:1" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0200, 0x0200, "SW2:2" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0400, 0x0400, "SW2:3" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x0800, 0x0800, "SW2:4" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x1000, 0x1000, "SW2:5" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x2000, 0x2000, "SW2:6" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x4000, 0x4000, "SW2:7" )
-	PORT_DIPUNKNOWN_DIPLOC( 0x8000, 0x8000, "SW2:8" )
-INPUT_PORTS_END
-
-/**************************************************************************
 
 
                                 Gfx Layouts
@@ -1956,37 +1696,12 @@ INPUT_PORTS_END
 
 **************************************************************************/
 
-static const uint32_t road_layout_xoffset[64] =
-{
-	STEP16(16*4*0,4),STEP16(16*4*1,4),
-	STEP16(16*4*2,4),STEP16(16*4*3,4)
-};
-
-/* Road: 64 x 1 x 4 */
-static const gfx_layout road_layout =
-{
-	64,1,
-	RGN_FRAC(1,1),
-	4,
-	{ STEP4(0,1) },
-	EXTENDED_XOFFS,
-	{ 0 },
-	64*1*4,
-	road_layout_xoffset,
-	nullptr
-};
-
 /**************************************************************************
                                 Big Run
 **************************************************************************/
 
 static GFXDECODE_START( gfx_bigrun )
-	//GFXDECODE_ENTRY( "scroll0", 0, gfx_8x8x4_packed_msb,   0x0e00/2 , 16 ) // Scroll 0
-	//GFXDECODE_ENTRY( "scroll1", 0, gfx_8x8x4_packed_msb,   0x1600/2 , 16 ) // Scroll 1
-	//GFXDECODE_ENTRY( "scroll2", 0, gfx_8x8x4_packed_msb,   0x3600/2 , 16 ) // Scroll 2
-	GFXDECODE_ENTRY( "sprites", 0, gfx_16x16x4_packed_msb, 0x2800/2 , 64 ) // [0] Sprites
-	GFXDECODE_ENTRY( "road0", 0, road_layout,0x2000/2 , 64 ) // [1] Road 0
-	GFXDECODE_ENTRY( "road1", 0, road_layout,0x1800/2 , 64 ) // [2] Road 1
+	GFXDECODE_ENTRY( "sprites", 0, gfx_16x16x4_packed_msb, 0x2800 / 2, 64 ) // [0] Sprites
 GFXDECODE_END
 
 /**************************************************************************
@@ -1994,37 +1709,7 @@ GFXDECODE_END
 **************************************************************************/
 
 static GFXDECODE_START( gfx_cischeat )
-	//GFXDECODE_ENTRY( "scroll0", 0, gfx_8x8x4_packed_msb,   0x1c00/2, 32  ) // Scroll 0
-	//GFXDECODE_ENTRY( "scroll1", 0, gfx_8x8x4_packed_msb,   0x2c00/2, 32  ) // Scroll 1
-	//GFXDECODE_ENTRY( "scroll2", 0, gfx_8x8x4_packed_msb,   0x6c00/2, 32  ) // Scroll 2
-	GFXDECODE_ENTRY( "sprites", 0, gfx_16x16x4_packed_msb, 0x5000/2, 128 ) // [0] Sprites
-	GFXDECODE_ENTRY( "road0", 0, road_layout,0x3800/2, 64  ) // [1] Road 0
-	GFXDECODE_ENTRY( "road1", 0, road_layout,0x4800/2, 64  ) // [2] Road 1
-GFXDECODE_END
-
-/**************************************************************************
-                            F1 GrandPrix Star
-**************************************************************************/
-
-static GFXDECODE_START( gfx_f1gpstar )
-	//GFXDECODE_ENTRY( "scroll0", 0, gfx_8x8x4_packed_msb,   0x1e00/2, 16  ) // Scroll 0
-	//GFXDECODE_ENTRY( "scroll1", 0, gfx_8x8x4_packed_msb,   0x2e00/2, 16  ) // Scroll 1
-	//GFXDECODE_ENTRY( "scroll2", 0, gfx_8x8x4_packed_msb,   0x6e00/2, 16  ) // Scroll 2
-	GFXDECODE_ENTRY( "sprites", 0, gfx_16x16x4_packed_msb, 0x5000/2, 128 ) // [0] Sprites
-	GFXDECODE_ENTRY( "road0", 0, road_layout,0x3800/2, 64  ) // [1] Road 0
-	GFXDECODE_ENTRY( "road1", 0, road_layout,0x4800/2, 64  ) // [2] Road 1
-GFXDECODE_END
-
-/**************************************************************************
-                                Scud Hammer
-**************************************************************************/
-
-static GFXDECODE_START( gfx_scudhamm )
-	//GFXDECODE_ENTRY( "scroll0", 0, gfx_8x8x4_packed_msb,               0x1e00/2, 16  )   // Scroll 0
-	//GFXDECODE_ENTRY( "scroll0", 0, gfx_8x8x4_packed_msb,               0x0000/2, 16  )   // UNUSED
-	//GFXDECODE_ENTRY( "scroll2", 0, gfx_8x8x4_packed_msb,               0x4e00/2, 16  )   // Scroll 2
-	GFXDECODE_ENTRY( "sprites", 0, gfx_8x8x4_col_2x2_group_packed_msb, 0x3000/2, 128 )   // [0] sprites
-	// No Road Layers
+	GFXDECODE_ENTRY( "sprites", 0, gfx_16x16x4_packed_msb, 0x5000 / 2, 128 ) // [0] Sprites
 GFXDECODE_END
 
 
@@ -2050,45 +1735,44 @@ GFXDECODE_END
 //         something still triggers it somehow?
 TIMER_DEVICE_CALLBACK_MEMBER(cischeat_state::bigrun_scanline)
 {
-	int scanline = param;
+	const int scanline = param;
 
-	if(m_screen->frame_number() & 1)
+	if (m_screen->frame_number() & 1)
 	{
-		if(scanline == 240)
-			m_cpu1->set_input_line(1, HOLD_LINE);
+		if (scanline == 240)
+			m_maincpu->set_input_line(1, HOLD_LINE);
 
 		return;
 	}
 
-	if(scanline == 240) // vblank-out irq
-		m_cpu1->set_input_line(4, HOLD_LINE);
+	if (scanline == 240) // vblank-out irq
+		m_maincpu->set_input_line(4, HOLD_LINE);
 
-	if(scanline == 0)
-		m_cpu1->set_input_line(2, HOLD_LINE);
+	if (scanline == 0)
+		m_maincpu->set_input_line(2, HOLD_LINE);
 }
 
 void cischeat_state::sound_irq(int state)
 {
-	if(state)
+	if (state)
 		m_soundcpu->set_input_line(4, HOLD_LINE);
 }
-
 
 
 void cischeat_state::bigrun(machine_config &config)
 {
 	/* basic machine hardware */
-	M68000(config, m_cpu1, 10000000);
-	m_cpu1->set_addrmap(AS_PROGRAM, &cischeat_state::bigrun_map);
+	M68000(config, m_maincpu, 10000000);
+	m_maincpu->set_addrmap(AS_PROGRAM, &cischeat_state::bigrun_map);
 	TIMER(config, "scantimer").configure_scanline(FUNC(cischeat_state::bigrun_scanline), "screen", 0, 1);
 
-	M68000(config, m_cpu2, 10000000);
-	m_cpu2->set_addrmap(AS_PROGRAM, &cischeat_state::bigrun_map2);
-	m_cpu2->set_vblank_int("screen", FUNC(cischeat_state::irq4_line_hold));
+	M68000(config, m_road1cpu, 10000000);
+	m_road1cpu->set_addrmap(AS_PROGRAM, &cischeat_state::bigrun_map2);
+	m_road1cpu->set_vblank_int("screen", FUNC(cischeat_state::irq4_line_hold));
 
-	M68000(config, m_cpu3, 10000000);
-	m_cpu3->set_addrmap(AS_PROGRAM, &cischeat_state::bigrun_map3);
-	m_cpu3->set_vblank_int("screen", FUNC(cischeat_state::irq4_line_hold));
+	M68000(config, m_road2cpu, 10000000);
+	m_road2cpu->set_addrmap(AS_PROGRAM, &cischeat_state::bigrun_map3);
+	m_road2cpu->set_vblank_int("screen", FUNC(cischeat_state::irq4_line_hold));
 
 	M68000(config, m_soundcpu, 6000000);
 	m_soundcpu->set_addrmap(AS_PROGRAM, &cischeat_state::bigrun_sound_map);
@@ -2107,7 +1791,8 @@ void cischeat_state::bigrun(machine_config &config)
 	m_screen->set_screen_update(FUNC(cischeat_state::screen_update_bigrun));
 	m_screen->set_palette(m_palette);
 
-	GFXDECODE(config, m_gfxdecode, m_palette, gfx_bigrun);
+	JALECO_ZOOMSPR_BIGRUN(config, m_spritegen, m_palette, gfx_bigrun);
+
 	PALETTE(config, m_palette, palette_device::BLACK).set_format(palette_device::RRRRGGGGBBBBRGBx, 0x4000/2);
 	m_palette->enable_shadows();
 
@@ -2115,11 +1800,14 @@ void cischeat_state::bigrun(machine_config &config)
 	MEGASYS1_TILEMAP(config, m_tmap[1], m_palette, 0x1600/2);
 	MEGASYS1_TILEMAP(config, m_tmap[2], m_palette, 0x3600/2);
 
+	JALECO_ROAD(config, m_road[0], m_palette, 0x2000/2);
+	JALECO_ROAD(config, m_road[1], m_palette, 0x1800/2);
+
 	/* sound hardware */
 	SPEAKER(config, "speaker", 2).front();
 
-	GENERIC_LATCH_16(config, m_soundlatch);
-	GENERIC_LATCH_16(config, m_soundlatch2);
+	GENERIC_LATCH_16(config, m_soundlatch[0]);
+	GENERIC_LATCH_16(config, m_soundlatch[1]);
 
 	// TODO: all sound frequencies unverified (assume same as Mega System 1)
 	ym2151_device &ymsnd(YM2151(config, "ymsnd", 7000000/2));
@@ -2127,13 +1815,13 @@ void cischeat_state::bigrun(machine_config &config)
 	ymsnd.add_route(0, "speaker", 0.50, 0);
 	ymsnd.add_route(1, "speaker", 0.50, 1);
 
-	OKIM6295(config, m_oki1, 4000000, okim6295_device::PIN7_HIGH); // clock frequency & pin 7 not verified
-	m_oki1->add_route(ALL_OUTPUTS, "speaker", 0.25, 0);
-	m_oki1->add_route(ALL_OUTPUTS, "speaker", 0.25, 1);
+	OKIM6295(config, m_oki[0], 4000000, okim6295_device::PIN7_HIGH); // clock frequency & pin 7 not verified
+	m_oki[0]->add_route(ALL_OUTPUTS, "speaker", 0.25, 0);
+	m_oki[0]->add_route(ALL_OUTPUTS, "speaker", 0.25, 1);
 
-	OKIM6295(config, m_oki2, 4000000, okim6295_device::PIN7_HIGH); // clock frequency & pin 7 not verified
-	m_oki2->add_route(ALL_OUTPUTS, "speaker", 0.25, 0);
-	m_oki2->add_route(ALL_OUTPUTS, "speaker", 0.25, 1);
+	OKIM6295(config, m_oki[1], 4000000, okim6295_device::PIN7_HIGH); // clock frequency & pin 7 not verified
+	m_oki[1]->add_route(ALL_OUTPUTS, "speaker", 0.25, 0);
+	m_oki[1]->add_route(ALL_OUTPUTS, "speaker", 0.25, 1);
 }
 
 void cischeat_state::bigrun_d65006(machine_config &config)
@@ -2149,19 +1837,24 @@ void cischeat_state::cischeat(machine_config &config)
 	bigrun(config);
 
 	/* basic machine hardware */
-	m_cpu1->set_addrmap(AS_PROGRAM, &cischeat_state::cischeat_map);
+	m_maincpu->set_addrmap(AS_PROGRAM, &cischeat_state::cischeat_map);
 
-	m_cpu2->set_addrmap(AS_PROGRAM, &cischeat_state::cischeat_map2);
+	m_road1cpu->set_addrmap(AS_PROGRAM, &cischeat_state::cischeat_map2);
 
-	m_cpu3->set_addrmap(AS_PROGRAM, &cischeat_state::cischeat_map3);
+	m_road2cpu->set_addrmap(AS_PROGRAM, &cischeat_state::cischeat_map3);
 
 	m_soundcpu->set_addrmap(AS_PROGRAM, &cischeat_state::cischeat_sound_map);
+
+	MEGASYS1_GATEARRAY_GS88000(config, m_gatearray, 0);
+	m_gatearray->set_cpuspace_tag(m_soundcpu, AS_PROGRAM);
+	m_gatearray->set_cpuregion_tag("soundcpu");
 
 	/* video hardware */
 	m_screen->set_visarea(0, 256-1,  0+16, 256-16-8-1);
 	m_screen->set_screen_update(FUNC(cischeat_state::screen_update_cischeat));
 
-	m_gfxdecode->set_info(gfx_cischeat);
+	JALECO_ZOOMSPR(config.replace(), m_spritegen, m_palette, gfx_cischeat);
+
 	m_palette->set_format(palette_device::RRRRGGGGBBBBRGBx, 0x8000/2);
 
 	m_tmap[0]->set_colorbase(0x1c00/2);
@@ -2172,42 +1865,37 @@ void cischeat_state::cischeat(machine_config &config)
 
 	m_tmap[2]->set_colorbase(0x6c00/2);
 	m_tmap[2]->set_bits_per_color_code(5);
-}
 
-void cischeat_state::cischeat_gs88000(machine_config &config)
-{
-	cischeat(config);
-	MEGASYS1_GATEARRAY_GS88000(config, m_gatearray, 0);
-	m_gatearray->set_cpuspace_tag(m_soundcpu, AS_PROGRAM);
-	m_gatearray->set_cpuregion_tag("soundcpu");
+	m_road[0]->set_colorbase(0x3800/2);
+	m_road[1]->set_colorbase(0x4800/2);
 }
-
 
 void cischeat_state::f1gpstar(machine_config &config)
 {
 	bigrun(config);
 
 	/* basic machine hardware */
-	m_cpu1->set_clock(12000000);
-	m_cpu1->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstar_map);
+	m_maincpu->set_clock(12000000);
+	m_maincpu->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstar_map);
 
-	m_cpu2->set_clock(12000000);
-	m_cpu2->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstar_map2);
+	m_road1cpu->set_clock(12000000);
+	m_road1cpu->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstar_map2);
 
-	m_cpu3->set_clock(12000000);
-	m_cpu3->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstar_map3);
+	m_road2cpu->set_clock(12000000);
+	m_road2cpu->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstar_map3);
 
 	m_soundcpu->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstar_sound_map);
 
 	/* video hardware */
-	m_gfxdecode->set_info(gfx_f1gpstar);
+	JALECO_ZOOMSPR(config.replace(), m_spritegen, m_palette, gfx_cischeat);
 	m_palette->set_format(palette_device::RRRRGGGGBBBBRGBx, 0x8000/2);
 
 	m_tmap[0]->set_colorbase(0x1e00/2);
-
 	m_tmap[1]->set_colorbase(0x2e00/2);
-
 	m_tmap[2]->set_colorbase(0x6e00/2);
+
+	JALECO_ZOOM_ROAD(config.replace(), m_road[0], m_palette, 0x3800/2);
+	JALECO_ZOOM_ROAD(config.replace(), m_road[1], m_palette, 0x4800/2);
 
 	m_screen->set_screen_update(FUNC(cischeat_state::screen_update_f1gpstar));
 }
@@ -2219,12 +1907,12 @@ void cischeat_state::f1gpstr2(machine_config &config)
 
 	/* basic machine hardware */
 
-	m_cpu1->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstr2_map);
+	m_maincpu->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstr2_map);
 
 	m_soundcpu->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstr2_sound_map);
 
-	M68000(config, m_cpu5, 10000000);
-	m_cpu5->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstr2_io_map);
+	M68000(config, m_iocpu, 10000000);
+	m_iocpu->set_addrmap(AS_PROGRAM, &cischeat_state::f1gpstr2_io_map);
 
 	config.set_maximum_quantum(attotime::from_hz(12000));
 }
@@ -2233,181 +1921,7 @@ void cischeat_state::f1gpstr2(machine_config &config)
 void wildplt_state::wildplt(machine_config &config)
 {
 	f1gpstr2(config);
-	m_cpu1->set_addrmap(AS_PROGRAM, &wildplt_state::wildplt_map);
-}
-
-
-/**************************************************************************
-                                Scud Hammer
-**************************************************************************/
-
-/*
-    1, 5-7]     busy loop
-    2]          clr.w   $fc810.l + rte
-    3]          game
-    4]          == 3
-*/
-
-TIMER_DEVICE_CALLBACK_MEMBER(cischeat_state::scudhamm_scanline)
-{
-	int scanline = param;
-
-	if(m_screen->frame_number() & 1)
-		return;
-
-	if(scanline == 240) // vblank-out irq
-		m_maincpu->set_input_line(4, HOLD_LINE);
-
-	if(scanline == 16) // clears a flag, sprite end DMA?
-		m_maincpu->set_input_line(2, HOLD_LINE);
-}
-
-void cischeat_state::scudhamm(machine_config &config)
-{
-	/* basic machine hardware */
-	M68000(config, m_maincpu, 12000000);
-	m_maincpu->set_addrmap(AS_PROGRAM, &cischeat_state::scudhamm_map);
-	TIMER(config, "scantimer").configure_scanline(FUNC(cischeat_state::scudhamm_scanline), "screen", 0, 1);
-
-	adc0804_device &adc(ADC0804(config, "adc", 640000)); // unknown clock
-	adc.vin_callback().set(FUNC(cischeat_state::scudhamm_analog_r));
-
-	WATCHDOG_TIMER(config, m_watchdog);
-
-	/* video hardware */
-	SCREEN(config, m_screen, SCREEN_TYPE_RASTER);
-	m_screen->set_video_attributes(VIDEO_UPDATE_AFTER_VBLANK);
-	// measured values for Arm Champs II: VSync: 59.1784Hz, HSync: 15082.0 kHz
-	m_screen->set_raw(XTAL(12'000'000)/2,396,0,256,256,16,240);
-//  m_screen->set_refresh_hz(30); //TODO: wrong!
-//  m_screen->set_vblank_time(ATTOSECONDS_IN_USEC(2500 * 3) /* not accurate */);
-//  m_screen->set_size(256, 256);
-//  m_screen->set_visarea(0, 256-1, 0 +16, 256-1 -16);
-	m_screen->set_screen_update(FUNC(cischeat_state::screen_update_scudhamm));
-	m_screen->set_palette(m_palette);
-
-	GFXDECODE(config, m_gfxdecode, m_palette, gfx_scudhamm);
-	PALETTE(config, m_palette, palette_device::BLACK).set_format(palette_device::RRRRGGGGBBBBRGBx, 0x8000/2);
-	m_palette->enable_shadows();
-
-	MEGASYS1_TILEMAP(config, m_tmap[0], m_palette, 0x1e00/2);
-	MEGASYS1_TILEMAP(config, m_tmap[2], m_palette, 0x4e00/2);
-
-	/* sound hardware */
-	SPEAKER(config, "speaker", 2).front();
-
-	OKIM6295(config, m_oki1, 4000000/2, okim6295_device::PIN7_HIGH); // pin 7 not verified
-	m_oki1->add_route(ALL_OUTPUTS, "speaker", 0.5, 0);
-	m_oki1->add_route(ALL_OUTPUTS, "speaker", 0.5, 1);
-
-	OKIM6295(config, m_oki2, 4000000/2, okim6295_device::PIN7_HIGH); // pin 7 not verified
-	m_oki2->add_route(ALL_OUTPUTS, "speaker", 0.5, 0);
-	m_oki2->add_route(ALL_OUTPUTS, "speaker", 0.5, 1);
-}
-
-
-/**************************************************************************
-                            Arm Champs II
-**************************************************************************/
-
-// TODO: intentionally duplicate scudhamm_scanline fn
-// armchamp2_state shouldn't derive from scudhamm_state but share this as a common device,
-// also assuming it really is using the same irq config as well ...
-TIMER_DEVICE_CALLBACK_MEMBER(armchamp2_state::armchamp2_scanline)
-{
-	int scanline = param;
-
-	if(m_screen->frame_number() & 1)
-		return;
-
-	if(scanline == 240) // vblank-out irq
-		m_maincpu->set_input_line(4, HOLD_LINE);
-
-	if(scanline == 16) // does a bit more than scudhammer
-		m_maincpu->set_input_line(2, HOLD_LINE);
-}
-
-void armchamp2_state::armchmp2(machine_config &config)
-{
-	scudhamm(config);
-
-	/* basic machine hardware */
-	m_maincpu->set_addrmap(AS_PROGRAM, &armchamp2_state::armchmp2_map);
-	subdevice<timer_device>("scantimer")->set_callback(FUNC(armchamp2_state::armchamp2_scanline));
-	subdevice<adc0804_device>("adc")->vin_callback().set(FUNC(armchamp2_state::analog_r));
-}
-
-
-/**************************************************************************
-                                Captain Flag
-**************************************************************************/
-
-/*
-    1]          no-op
-    2]          copy text buffer to tilemap, sound (~50 scanlines at 30Hz)
-    3]          I/O, sound (~5 scanlines at 30Hz). Sets a flag to update text buffer and sprites
-    4-7]        rte
-*/
-
-TIMER_DEVICE_CALLBACK_MEMBER(captflag_state::captflag_scanline)
-{
-	int scanline = param;
-
-	if(scanline == 240) // vblank: draw screen
-		m_maincpu->set_input_line(2, HOLD_LINE);
-
-	if(scanline == 50)
-		m_maincpu->set_input_line(3, HOLD_LINE);
-}
-
-void captflag_state::captflag(machine_config &config)
-{
-	/* basic machine hardware */
-	M68000(config, m_maincpu, XTAL(24'000'000) / 2);  // TMP68000P-12
-	m_maincpu->set_addrmap(AS_PROGRAM, &captflag_state::captflag_map);
-	TIMER(config, "scantimer").configure_scanline(FUNC(captflag_state::captflag_scanline), "screen", 0, 1);
-
-	TICKET_DISPENSER(config, m_hopper, attotime::from_msec(2000));
-
-	WATCHDOG_TIMER(config, m_watchdog);
-	NVRAM(config, "nvram", nvram_device::DEFAULT_ALL_0);
-
-	/* video hardware */
-	SCREEN(config, m_screen, SCREEN_TYPE_RASTER);
-//  m_screen->set_video_attributes(VIDEO_UPDATE_AFTER_VBLANK);
-	m_screen->set_refresh_hz(30); //TODO: wrong!
-//  m_screen->set_vblank_time(ATTOSECONDS_IN_USEC(2500 * 3) /* not accurate */);
-	m_screen->set_size(256, 256);
-	m_screen->set_visarea(0, 256-1, 0 +16, 256-1 -16);
-	m_screen->set_screen_update(FUNC(captflag_state::screen_update_scudhamm));
-	m_screen->set_palette(m_palette);
-
-	GFXDECODE(config, m_gfxdecode, m_palette, gfx_scudhamm);
-	PALETTE(config, m_palette, palette_device::BLACK).set_format(palette_device::RRRRGGGGBBBBRGBx, 0x8000/2);
-	m_palette->enable_shadows();
-
-	MEGASYS1_TILEMAP(config, m_tmap[0], m_palette, 0x1e00/2);
-	MEGASYS1_TILEMAP(config, m_tmap[2], m_palette, 0x4e00/2);
-
-	// Motors
-	TIMER(config, m_motor_left).configure_generic(nullptr);
-	TIMER(config, m_motor_right).configure_generic(nullptr);
-
-	// Layout
-	config.set_default_layout(layout_captflag);
-
-	/* sound hardware */
-	SPEAKER(config, "speaker", 2).front();
-
-	OKIM6295(config, m_oki1, 4000000/2, okim6295_device::PIN7_HIGH); // pin 7 not verified
-	m_oki1->set_addrmap(0, &captflag_state::oki1_map);
-	m_oki1->add_route(ALL_OUTPUTS, "speaker", 0.5, 0);
-	m_oki1->add_route(ALL_OUTPUTS, "speaker", 0.5, 1);
-
-	OKIM6295(config, m_oki2, 4000000/2, okim6295_device::PIN7_HIGH); // pin 7 not verified
-	m_oki2->set_addrmap(0, &captflag_state::oki2_map);
-	m_oki2->add_route(ALL_OUTPUTS, "speaker", 0.5, 0);
-	m_oki2->add_route(ALL_OUTPUTS, "speaker", 0.5, 1);
+	m_maincpu->set_addrmap(AS_PROGRAM, &wildplt_state::wildplt_map);
 }
 
 
@@ -2432,24 +1946,22 @@ void captflag_state::captflag(machine_config &config)
 */
 void cischeat_state::cischeat_untangle_sprites(const char *region)
 {
-	uint8_t       *src = memregion(region)->base();
-	const uint8_t *end = src + memregion(region)->bytes();
+	u8       *src = memregion(region)->base();
+	const u8 *end = src + memregion(region)->bytes();
 
 	while (src < end)
 	{
-		uint8_t sprite[16*8];
-		int i;
+		u8 sprite[16 * 8];
 
-		for (i = 0; i < 16 ; i++)
+		for (int i = 0; i < 16 ; i++)
 		{
-			memcpy(&sprite[i*8+0], &src[i*4+0],    4);
-			memcpy(&sprite[i*8+4], &src[i*4+16*4], 4);
+			memcpy(&sprite[i * 8 + 0], &src[i * 4 + 0],      4);
+			memcpy(&sprite[i * 8 + 4], &src[i * 4 + 16 * 4], 4);
 		}
-		memcpy(src, sprite, 16*8);
-		src += 16*8;
+		memcpy(src, sprite, 16 * 8);
+		src += 16 * 8;
 	}
 }
-
 
 
 /***************************************************************************
@@ -2517,15 +2029,15 @@ BR8950c
 ***************************************************************************/
 
 ROM_START( bigrun )
-	ROM_REGION( 0x80000, "cpu1", 0 )
+	ROM_REGION( 0x80000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "br8950b.e1",  0x000000, 0x040000, CRC(bfb54a62) SHA1(49f78e162e8bc19a75c62029737acd665b9b124b) )
 	ROM_LOAD16_BYTE( "br8950b.e2",  0x000001, 0x040000, CRC(c0483e81) SHA1(e8fd8860191c7d8cb4dda44c69ce05cd58174a07) )
 
-	ROM_REGION( 0x40000, "cpu2", 0 )
+	ROM_REGION( 0x40000, "road1cpu", 0 )
 	ROM_LOAD16_BYTE( "br8952a.19", 0x000000, 0x020000, CRC(fcf6b70c) SHA1(34f7ca29241925251a043c63062596d1c220b730) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 	ROM_LOAD16_BYTE( "br8952a.20", 0x000001, 0x020000, CRC(c43d367b) SHA1(86242ba03cd172e95bf42eac369fb3acf42c7c45) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 
-	ROM_REGION( 0x40000, "cpu3", 0 )
+	ROM_REGION( 0x40000, "road2cpu", 0 )
 	ROM_LOAD16_BYTE( "br8952a.9",  0x000000, 0x020000, CRC(f803f0d9) SHA1(6d2e90e74bb15cd443d901ad0fc827117432daa6) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 	ROM_LOAD16_BYTE( "br8952a.10", 0x000001, 0x020000, CRC(8c0df287) SHA1(3a2e689e2d841e2089b18267a8462f1acc40ae99) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 
@@ -2602,15 +2114,15 @@ ROM_START( bigrun )
 ROM_END
 
 ROM_START( bigrunu )
-	ROM_REGION( 0x80000, "cpu1", 0 )
+	ROM_REGION( 0x80000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "br8950b.1",  0x000000, 0x040000, CRC(748d1525) SHA1(1e7d8d8bae07d4b81ca46640ffcbd501b5504008) )
 	ROM_LOAD16_BYTE( "br8950b.2",  0x000001, 0x040000, CRC(4164a542) SHA1(d889388784a9ecc502071ab7ccf05e8574d25cf1) )
 
-	ROM_REGION( 0x40000, "cpu2", 0 )
+	ROM_REGION( 0x40000, "road1cpu", 0 )
 	ROM_LOAD16_BYTE( "br8952a.19", 0x000000, 0x020000, CRC(fcf6b70c) SHA1(34f7ca29241925251a043c63062596d1c220b730) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 	ROM_LOAD16_BYTE( "br8952a.20", 0x000001, 0x020000, CRC(c43d367b) SHA1(86242ba03cd172e95bf42eac369fb3acf42c7c45) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 
-	ROM_REGION( 0x40000, "cpu3", 0 )
+	ROM_REGION( 0x40000, "road2cpu", 0 )
 	ROM_LOAD16_BYTE( "br8952a.9",  0x000000, 0x020000, CRC(f803f0d9) SHA1(6d2e90e74bb15cd443d901ad0fc827117432daa6) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 	ROM_LOAD16_BYTE( "br8952a.10", 0x000001, 0x020000, CRC(8c0df287) SHA1(3a2e689e2d841e2089b18267a8462f1acc40ae99) ) // 1xxxxxxxxxxxxxxxx = 0xFF
 
@@ -2732,7 +2244,7 @@ Video                        Sound
 |                                |      ([9][10][11]:82S129)
  --------------------------------
 
-DIP SW:8BITx2 , 4BITx1
+DIP SW:8BITx2, 4BITx1
 
 According to KLOV:
 
@@ -2743,15 +2255,15 @@ Sound:      Amplified Stereo (two channel)
 ***************************************************************************/
 
 ROM_START( cischeat )
-	ROM_REGION( 0x080000, "cpu1", 0 )
+	ROM_REGION( 0x080000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "ch9071v2.03", 0x000000, 0x040000, CRC(dd1bb26f) SHA1(2b9330b45edcc3291ad4ac935558c1f070ab5bd9) )
 	ROM_LOAD16_BYTE( "ch9071v2.01", 0x000001, 0x040000, CRC(7b65276a) SHA1(e0075b6d09da12ab7c84b888ffe65cd33ec7c6b6) )
 
-	ROM_REGION( 0x80000, "cpu2", 0 )
+	ROM_REGION( 0x80000, "road1cpu", 0 )
 	ROM_LOAD16_BYTE( "ch9073.01",  0x000000, 0x040000, CRC(ba331526) SHA1(bed5182c0d63a64cc44f9c127d1b8ce55a701184) )
 	ROM_LOAD16_BYTE( "ch9073.02",  0x000001, 0x040000, CRC(b45ff10f) SHA1(11fb2d9a3b5cc4f997816e0d3c9cf47bf749db17) )
 
-	ROM_REGION( 0x80000, "cpu3", 0 )
+	ROM_REGION( 0x80000, "road2cpu", 0 )
 	ROM_LOAD16_BYTE( "ch9073v1.03", 0x000000, 0x040000, CRC(bf1d1cbf) SHA1(0892ce5f35864393a6f899f02f811b8fdba03978) )
 	ROM_LOAD16_BYTE( "ch9073v1.04", 0x000001, 0x040000, CRC(1ec8a597) SHA1(311d4aa8bd92dd2eea0a64f881f64d19b7ba7d12) )
 
@@ -2762,8 +2274,8 @@ ROM_START( cischeat )
 	ROM_REGION16_BE( 0x100000, "user1", 0 ) /* second halves of program ROMs */
 	ROM_LOAD16_BYTE( "ch9071.04",   0x000000, 0x040000, CRC(7fb48cbc) SHA1(7f0442ce37b39e830fe8bcb8230cf7da2103059d) )  // cpu #1
 	ROM_LOAD16_BYTE( "ch9071.02",   0x000001, 0x040000, CRC(a5d0f4dc) SHA1(2e7aaa915e27ab31e38ca6759301ffe33a12b427) )
-	ROM_COPY( "cpu2", 0x40000, 0x80000, 0x40000 )
-	ROM_COPY( "cpu3", 0x40000, 0xc0000, 0x40000 )
+	ROM_COPY( "road1cpu", 0x40000, 0x80000, 0x40000 )
+	ROM_COPY( "road2cpu", 0x40000, 0xc0000, 0x40000 )
 
 	ROM_REGION( 0x040000, "scroll0", 0 )
 	ROM_LOAD( "ch9071.a14",  0x000000, 0x040000, CRC(7a6d147f) SHA1(8f52e012d9699311c2a2409130c6200c6d2e1c51) ) // scroll 0
@@ -2933,16 +2445,16 @@ GFX & Misc       - GS90015-02 (100 pin PQFP),  uses ROM 90015-31-R56
 ***************************************************************************/
 
 ROM_START( f1gpstar )
-	ROM_REGION( 0x100000, "cpu1", 0 )
+	ROM_REGION( 0x100000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "gp9188a27ver4.0.bin", 0x000000, 0x040000, CRC(47224c65) SHA1(1b09fe93b1896cf0706d4914ef956480c1ca231f) )
 	ROM_LOAD16_BYTE( "gp9188a22ver4.0.bin", 0x000001, 0x040000, CRC(05ca6410) SHA1(3b8bfd1702c43c398bffda6931e5575110cc5962) )
 
-	ROM_REGION( 0x80000, "cpu2", 0 )
+	ROM_REGION( 0x80000, "road1cpu", 0 )
 	/* Should Use ROMs: 90015-01.W06, 90015-02.W07, 90015-03.W08, 90015-04.W09 */
 	ROM_LOAD16_BYTE( "9188a-16.v10",  0x000000, 0x020000, CRC(ef0f7ca9) SHA1(98ad687fdab67dd9f54b50cf21fd10ac34b61e7a) )
 	ROM_LOAD16_BYTE( "9188a-11.v10",  0x000001, 0x020000, CRC(de292ea3) SHA1(04ed19045edb4edfff2b8fedac37c4a3352dfa76) )
 
-	ROM_REGION( 0x80000, "cpu3", 0 )
+	ROM_REGION( 0x80000, "road2cpu", 0 )
 	/* Should Use ROMs: 90015-01.W06, 90015-02.W07, 90015-03.W08, 90015-04.W09 */
 	ROM_LOAD16_BYTE( "9188a-6.v10",  0x000000, 0x020000, CRC(18ba0340) SHA1(e46e10a350f18cf3a46c0d3a0cb08fc369fced6d) )
 	ROM_LOAD16_BYTE( "9188a-1.v10",  0x000001, 0x020000, CRC(109d2913) SHA1(e117556481e801d51b8526a143bc202dda222f7f) )
@@ -3017,16 +2529,16 @@ ROM_START( f1gpstar )
 ROM_END
 
 ROM_START( f1gpstar3 )
-	ROM_REGION( 0x100000, "cpu1", 0 )
+	ROM_REGION( 0x100000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "gp9188a27ver3.0.ic125", 0x000000, 0x040000, CRC(e3ab0085) SHA1(c2de842f427d5e982a42467ca7354b728f63bf88) )
 	ROM_LOAD16_BYTE( "gp9188a22ver3.0.ic92",  0x000001, 0x040000, CRC(7022061b) SHA1(241261f79072f09dbc325185e8e5b039c592a491) )
 
-	ROM_REGION( 0x80000, "cpu2", 0 )
+	ROM_REGION( 0x80000, "road1cpu", 0 )
 	/* Should Use ROMs: 90015-01.W06, 90015-02.W07, 90015-03.W08, 90015-04.W09 */
 	ROM_LOAD16_BYTE( "9188a-16.v10",  0x000000, 0x020000, CRC(ef0f7ca9) SHA1(98ad687fdab67dd9f54b50cf21fd10ac34b61e7a) )
 	ROM_LOAD16_BYTE( "9188a-11.v10",  0x000001, 0x020000, CRC(de292ea3) SHA1(04ed19045edb4edfff2b8fedac37c4a3352dfa76) )
 
-	ROM_REGION( 0x80000, "cpu3", 0 )
+	ROM_REGION( 0x80000, "road2cpu", 0 )
 	/* Should Use ROMs: 90015-01.W06, 90015-02.W07, 90015-03.W08, 90015-04.W09 */
 	ROM_LOAD16_BYTE( "9188a-6.v10",  0x000000, 0x020000, CRC(18ba0340) SHA1(e46e10a350f18cf3a46c0d3a0cb08fc369fced6d) )
 	ROM_LOAD16_BYTE( "9188a-1.v10",  0x000001, 0x020000, CRC(109d2913) SHA1(e117556481e801d51b8526a143bc202dda222f7f) )
@@ -3101,16 +2613,16 @@ ROM_START( f1gpstar3 )
 ROM_END
 
 ROM_START( f1gpstar2 )
-	ROM_REGION( 0x100000, "cpu1", 0 )
+	ROM_REGION( 0x100000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "9188a-27.v20", 0x000000, 0x040000, CRC(0a9d3896) SHA1(5e3332a1b779dead1e4f9ef274a2f168721db0ed) )
 	ROM_LOAD16_BYTE( "9188a-22.v20", 0x000001, 0x040000, CRC(de15c9ca) SHA1(f356b02ca66b7e8ab0293e6e28fcd3f7996c80c8) )
 
-	ROM_REGION( 0x80000, "cpu2", 0 )
+	ROM_REGION( 0x80000, "road1cpu", 0 )
 	/* Should Use ROMs: 90015-01.W06, 90015-02.W07, 90015-03.W08, 90015-04.W09 */
 	ROM_LOAD16_BYTE( "9188a-16.v10",  0x000000, 0x020000, CRC(ef0f7ca9) SHA1(98ad687fdab67dd9f54b50cf21fd10ac34b61e7a) )
 	ROM_LOAD16_BYTE( "9188a-11.v10",  0x000001, 0x020000, CRC(de292ea3) SHA1(04ed19045edb4edfff2b8fedac37c4a3352dfa76) )
 
-	ROM_REGION( 0x80000, "cpu3", 0 )
+	ROM_REGION( 0x80000, "road2cpu", 0 )
 	/* Should Use ROMs: 90015-01.W06, 90015-02.W07, 90015-03.W08, 90015-04.W09 */
 	ROM_LOAD16_BYTE( "9188a-6.v10",  0x000000, 0x020000, CRC(18ba0340) SHA1(e46e10a350f18cf3a46c0d3a0cb08fc369fced6d) )
 	ROM_LOAD16_BYTE( "9188a-1.v10",  0x000001, 0x020000, CRC(109d2913) SHA1(e117556481e801d51b8526a143bc202dda222f7f) )
@@ -3268,7 +2780,7 @@ Main processor(A)  - 68000
                    - Processor RAM 2x LH52250AD-90L (32kx8 SRAM)
 
 Slave processor(B) - 68000
-                   - program ROMs: GP-9188A_11_Ver1-0.bin & , GP-9188A_16_Ver1-0.bin
+                   - program ROMs: GP-9188A_11_Ver1-0.bin &, GP-9188A_16_Ver1-0.bin
                    - Processor RAM 2x LH5168D-10L (8kx8 SRAM)
                    - CS90015-04 (64 pin PQFP)  + 2x MCM2018AN45 (2kx8 SRAM)
                    - GS90015-05 (100 pin PQFP) + 2x MCM2018AN45 (2kx8 SRAM)
@@ -3314,15 +2826,15 @@ GFX & Misc         - GS90015-02 (100 pin PQFP)
 ***************************************************************************/
 
 ROM_START( wildplt )
-	ROM_REGION( 0x100000, "cpu1", 0 )
+	ROM_REGION( 0x100000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "gp-9188a_27_ver1-4.bin", 0x000000, 0x40000, CRC(4754f023) SHA1(32865343d633fef7d3e081d52b423fbf4e5f2053) )
 	ROM_LOAD16_BYTE( "gp-9188a_22_ver1-4.bin", 0x000001, 0x40000, CRC(9e099111) SHA1(11b17d77dfe6315105e8297b8f5f7c8d93b18ac0) )
 
-	ROM_REGION( 0x80000, "cpu2", 0 )
+	ROM_REGION( 0x80000, "road1cpu", 0 )
 	ROM_LOAD16_BYTE( "gp-9188a_16_ver1-0.bin", 0x000000, 0x20000, CRC(75e48920) SHA1(2344c1f394aa1199169e75b2f998e89648a39100) )
 	ROM_LOAD16_BYTE( "gp-9188a_11_ver1-0.bin", 0x000001, 0x20000, CRC(8174e65b) SHA1(2c35ae0d04fc4a075c297fe69c429f57457348e1) )
 
-	ROM_REGION( 0x80000, "cpu3", 0 )
+	ROM_REGION( 0x80000, "road2cpu", 0 )
 	ROM_LOAD16_BYTE( "gp-9188a_06_ver1-0.bin", 0x000000, 0x20000, CRC(99139696) SHA1(4f8c96012a8214ed952aa44d65ff92e9b3adcf73) )
 	ROM_LOAD16_BYTE( "gp-9188a_01_ver1-0.bin", 0x000001, 0x20000, CRC(2c3d7ec4) SHA1(30215b2ea106aa3c8069fef9411e5f41be5645f6) )
 
@@ -3330,7 +2842,7 @@ ROM_START( wildplt )
 	ROM_LOAD16_BYTE( "wp-92116_3_ver1-1.bin", 0x000000, 0x20000, CRC(51dd594b) SHA1(3fbec13c1d2942797267820157180d374b38347c) )
 	ROM_LOAD16_BYTE( "wp-92116_4_ver1-1.bin", 0x000001, 0x20000, CRC(9ba3cb7b) SHA1(6f55e2675bde2756a2996e6ab6aaef4c9d3f4724) )
 
-	ROM_REGION( 0x80000, "cpu5", 0 )
+	ROM_REGION( 0x80000, "iocpu", 0 )
 	ROM_LOAD16_BYTE( "wp-92116_5_ver1-0.bin", 0x000000, 0x20000, CRC(8952e07c) SHA1(9025954c09c20dfe910a83283cfe010a5e898f38) )
 	ROM_LOAD16_BYTE( "wp-92116_6_ver1-0.bin", 0x000001, 0x20000, CRC(2c8108c2) SHA1(759ab884a09669c8cb532f72aa14a3df2806c5f2) )
 
@@ -3522,15 +3034,15 @@ PROMs :
 ***************************************************************************/
 
 ROM_START( f1gpstr2 )
-	ROM_REGION( 0x100000, "cpu1", 0 )
+	ROM_REGION( 0x100000, "maincpu", 0 )
 	ROM_LOAD16_BYTE( "9188a-27.125", 0x000000, 0x040000, CRC(ee60b894) SHA1(92247cd3b0e3fb2ed0ad27062d1cc13dadb21465) )
-	ROM_LOAD16_BYTE( "9188a-22.92" , 0x000001, 0x040000, CRC(f229332b) SHA1(f7037515d77a1f42ce555b8baa15075d7009a5c6) )
+	ROM_LOAD16_BYTE( "9188a-22.92",  0x000001, 0x040000, CRC(f229332b) SHA1(f7037515d77a1f42ce555b8baa15075d7009a5c6) )
 
-	ROM_REGION( 0x80000, "cpu2", 0 )
+	ROM_REGION( 0x80000, "road1cpu", 0 )
 	ROM_LOAD16_BYTE( "9188a-16.70",  0x000000, 0x020000, CRC(3b3d57a0) SHA1(2ad367fc938c385160a014b994fe45d1791d4ddb) )
 	ROM_LOAD16_BYTE( "9188a-11.46",  0x000001, 0x020000, CRC(4d2afddf) SHA1(108ddbb6a008d390e43d88efd10d99b599c1d75d) )
 
-	ROM_REGION( 0x80000, "cpu3", 0 )
+	ROM_REGION( 0x80000, "road2cpu", 0 )
 	ROM_LOAD16_BYTE( "9188a-6.27",  0x000000, 0x020000, CRC(68faa23b) SHA1(b18c407cc1273295ec9fc30af191fccd94447ae8) )
 	ROM_LOAD16_BYTE( "9188a-1.2",   0x000001, 0x020000, CRC(e4d83b4f) SHA1(4022521e43f6361c5e04a604f8a7fa1e60a2ac99) )
 
@@ -3538,7 +3050,7 @@ ROM_START( f1gpstr2 )
 	ROM_LOAD16_BYTE( "92116-3.37", 0x000000, 0x020000, CRC(2a541095) SHA1(934ef9b6bbe3f6e2e2649dde5671547b955ffc7c) )
 	ROM_LOAD16_BYTE( "92116-4.38", 0x000001, 0x020000, CRC(70da1825) SHA1(7f2077e9b40d5acf4da3f6bcc5d7b92d6d08f861) )
 
-	ROM_REGION( 0x80000, "cpu5", 0 )
+	ROM_REGION( 0x80000, "iocpu", 0 )
 	ROM_LOAD16_BYTE( "92116-5.116",  0x000000, 0x040000, CRC(da16db49) SHA1(a07fb706b0c93a83148a9fdaaca1bc5414bfe286) )
 	ROM_LOAD16_BYTE( "92116-6.117",  0x000001, 0x040000, CRC(1f1e147a) SHA1(ebedcdad9cfda8fa3b5c2653232209da5be237e1) )
 
@@ -3606,521 +3118,7 @@ ROM_START( f1gpstr2 )
 	ROM_LOAD( "pr90015b.153", 0x000000, 0x000100, CRC(be240dac) SHA1(6203b73c1a5e09e525380a78b555c3818929d5eb) )    // FIXED BITS (000xxxxx000xxxx1)
 ROM_END
 
-
-/***************************************************************************
-
-                                Scud Hammer
-
-CF-92128B:
-
-                                                      GS9001501
- 2-H 2-L  6295            62256 62256
- 1-H 1-L  6295  68000-12  3     4       6  GS90015-02 8464 8464 GS600406
-
-                    24MHz               5  GS90015-02 8464 8464 GS900406
-
-                                                       7C199
-                                                       7C199 GS90015-03
-
-GP-9189:
-
- 1     2      62256                            62256
- 3     4      62256    GS90015-06 GS90015-06   62256
- 5     6      62256                            62256
- 7     8      62256    GS90015-03 GS90015-03   62256
- 9     10
-
-                  GS90015-08            GS90015-07 GS90015-10
-
-          GS90015-11
-
-
-                      MR90015-35
-                      MR90015-35              GS90015-09
-
-***************************************************************************/
-
-ROM_START( scudhamm )
-	ROM_REGION( 0x080000, "maincpu", 0 )        /* Main CPU Code */
-	ROM_LOAD16_BYTE( "scud_hammer_cf-92128b-3_ver.1.4.bin", 0x000000, 0x040000, CRC(4747370e) SHA1(d822e949d444d3fe4b95edb58fa02b6b7a1a64e6) )
-	ROM_LOAD16_BYTE( "scud_hammer_cf-92128b-4_ver.1.4.bin", 0x000001, 0x040000, CRC(7a04955c) SHA1(57bbb3d481ebe9ae99d8ffae21984d53bdfe574a) )
-
-	ROM_REGION( 0x080000, "scroll0", 0 ) /* Scroll 0 */
-	ROM_LOAD( "5", 0x000000, 0x080000, CRC(714c115e) SHA1(c3e88b3972e3926f37968f3e84b932e1ac177142) )
-
-//  ROM_REGION( 0x080000, "scroll1", 0 ) /* Scroll 1 */
-//  UNUSED
-
-	ROM_REGION( 0x020000, "scroll2", 0 ) /* Scroll 2 */
-	ROM_LOAD( "6", 0x000000, 0x020000, CRC(b39aab63) SHA1(88275cce8b1323b2d835390a8fc2380b90d50d95) ) // 1xxxxxxxxxxxxxxxx = 0xFF
-
-	ROM_REGION( 0x500000, "sprites", 0 ) /* Sprites */
-	ROM_LOAD16_BYTE( "1.bot",  0x000000, 0x080000, CRC(46450d73) SHA1(c9acdf1cef760e5194c346d721e859c61afbfce6) )
-	ROM_LOAD16_BYTE( "2.bot",  0x000001, 0x080000, CRC(fb7b66dd) SHA1(ad6bbae4fa72f957e5c0fc7bf6199ac45f837dac) )
-	ROM_LOAD16_BYTE( "3.bot",  0x100000, 0x080000, CRC(7d45960b) SHA1(abf59cf85f28c90d4c08e3a1e5408a9a700071cc) )
-	ROM_LOAD16_BYTE( "4.bot",  0x100001, 0x080000, CRC(393b6a22) SHA1(0d002a8c09de2fb8aaa7f5f020badc6fc096fa41) )
-	ROM_LOAD16_BYTE( "5.bot",  0x200000, 0x080000, CRC(7a3c33ad) SHA1(fe0e3722e15919ae3acfeeacae57716aae43647c) )
-	ROM_LOAD16_BYTE( "6.bot",  0x200001, 0x080000, CRC(d19c4bf7) SHA1(b8aa21920d5a02f10a7ae65ade8a0a88ad23f373) )
-	ROM_LOAD16_BYTE( "7.bot",  0x300000, 0x080000, CRC(9e5edf59) SHA1(fcd4136e39d40bcce365153c96e06181a24a480a) )
-	ROM_LOAD16_BYTE( "8.bot",  0x300001, 0x080000, CRC(4980051e) SHA1(10de91239b5b4dab8e7fa4bf51d93356c5111ddf) )
-	ROM_LOAD16_BYTE( "9.bot",  0x400000, 0x080000, CRC(c1b301f1) SHA1(776b9889703d73afc4fb0ff77498b98c943246d3) )
-	ROM_LOAD16_BYTE( "10.bot", 0x400001, 0x080000, CRC(dab4528f) SHA1(f5ddc37a2d106d5438ad1b7d23a2bbbce07f2c89) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "2.l",  0x000000, 0x080000, CRC(889311da) SHA1(fcaee3e6c98a784cfde06fc2e0e8f5abbfb4df6c) )
-	ROM_LOAD( "2.h",  0x080000, 0x080000, CRC(347928fc) SHA1(36903c38b9f13594de40dfc697326327c7010d65) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "1.l",  0x000000, 0x080000, CRC(3c94aa90) SHA1(f9278fec9d93dac0309f30e35c727bd481f347d4) )
-	ROM_LOAD( "1.h",  0x080000, 0x080000, CRC(5caee787) SHA1(267f4d3c28e71e53180a5d0ff27a6555ac6fa4a0) )    // 1xxxxxxxxxxxxxxxxxx = 0xFF
-ROM_END
-
-ROM_START( scudhamma )
-	ROM_REGION( 0x080000, "maincpu", 0 )        /* Main CPU Code */
-	ROM_LOAD16_BYTE( "3", 0x000000, 0x040000, CRC(a908e7bd) SHA1(be0a8f959ab5c19122eee6c3def6137f37f1a9c6) )
-	ROM_LOAD16_BYTE( "4", 0x000001, 0x040000, CRC(981c8b02) SHA1(db6c8993bf1c3993ab31dd649022ab76169975e1) )
-
-	ROM_REGION( 0x080000, "scroll0", 0 ) /* Scroll 0 */
-	ROM_LOAD( "5", 0x000000, 0x080000, CRC(714c115e) SHA1(c3e88b3972e3926f37968f3e84b932e1ac177142) )
-
-//  ROM_REGION( 0x080000, "scroll1", 0 ) /* Scroll 1 */
-//  UNUSED
-
-	ROM_REGION( 0x020000, "scroll2", 0 ) /* Scroll 2 */
-	ROM_LOAD( "6", 0x000000, 0x020000, CRC(b39aab63) SHA1(88275cce8b1323b2d835390a8fc2380b90d50d95) ) // 1xxxxxxxxxxxxxxxx = 0xFF
-
-	ROM_REGION( 0x500000, "sprites", 0 ) /* Sprites */
-	ROM_LOAD16_BYTE( "1.bot",  0x000000, 0x080000, CRC(46450d73) SHA1(c9acdf1cef760e5194c346d721e859c61afbfce6) )
-	ROM_LOAD16_BYTE( "2.bot",  0x000001, 0x080000, CRC(fb7b66dd) SHA1(ad6bbae4fa72f957e5c0fc7bf6199ac45f837dac) )
-	ROM_LOAD16_BYTE( "3.bot",  0x100000, 0x080000, CRC(7d45960b) SHA1(abf59cf85f28c90d4c08e3a1e5408a9a700071cc) )
-	ROM_LOAD16_BYTE( "4.bot",  0x100001, 0x080000, CRC(393b6a22) SHA1(0d002a8c09de2fb8aaa7f5f020badc6fc096fa41) )
-	ROM_LOAD16_BYTE( "5.bot",  0x200000, 0x080000, CRC(7a3c33ad) SHA1(fe0e3722e15919ae3acfeeacae57716aae43647c) )
-	ROM_LOAD16_BYTE( "6.bot",  0x200001, 0x080000, CRC(d19c4bf7) SHA1(b8aa21920d5a02f10a7ae65ade8a0a88ad23f373) )
-	ROM_LOAD16_BYTE( "7.bot",  0x300000, 0x080000, CRC(9e5edf59) SHA1(fcd4136e39d40bcce365153c96e06181a24a480a) )
-	ROM_LOAD16_BYTE( "8.bot",  0x300001, 0x080000, CRC(4980051e) SHA1(10de91239b5b4dab8e7fa4bf51d93356c5111ddf) )
-	ROM_LOAD16_BYTE( "9.bot",  0x400000, 0x080000, CRC(c1b301f1) SHA1(776b9889703d73afc4fb0ff77498b98c943246d3) )
-	ROM_LOAD16_BYTE( "10.bot", 0x400001, 0x080000, CRC(dab4528f) SHA1(f5ddc37a2d106d5438ad1b7d23a2bbbce07f2c89) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "2.l",  0x000000, 0x080000, CRC(889311da) SHA1(fcaee3e6c98a784cfde06fc2e0e8f5abbfb4df6c) )
-	ROM_LOAD( "2.h",  0x080000, 0x080000, CRC(347928fc) SHA1(36903c38b9f13594de40dfc697326327c7010d65) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "1.l",  0x000000, 0x080000, CRC(3c94aa90) SHA1(f9278fec9d93dac0309f30e35c727bd481f347d4) )
-	ROM_LOAD( "1.h",  0x080000, 0x080000, CRC(5caee787) SHA1(267f4d3c28e71e53180a5d0ff27a6555ac6fa4a0) )    // 1xxxxxxxxxxxxxxxxxx = 0xFF
-ROM_END
-
-
-/***************************************************************************
-
-Arm Champs II
-Jaleco, 1992
-
-An arm wrestling game from Jaleco. Game is part mechanical, part video. The video part
-is shown on a vertical screen, being your opponent you are wrestling. The mechanical
-part is an arm you force against to pin your opponent's 'arm'. The mechanical arm is powered
-by a motor. The arm probably has a 5k potentiometer connected to it to give feedback to the
-hardware about the position of the mechanical arm.
-The opponents get progressively harder and harder, it's almost impossible to beat the final
-few opponents unless you have a few friends handy to swing on the arm ;-))
-The hardware appears to be similar to Jaleco's 'Grand Prix Star' etc, using many of
-the same custom IC's, and even some of the same ROMs.
-
-
-PCB Layouts
------------
-
-(top board)
-EB91042-20048-3
-AC-91106B
-|---------------------------------------------------------------------------|
-| uPC1318                                         GS9001501   TL7705        |
-|          |-------------|                                                  |
-|POWER     |   68000-12  |                                  PR88004Q_8.IC102|
-|          |-------------|                                                  |
-|                         24MHz                        AC91106_VER1.2_7.IC99|
-|                                                                           |
-|                                                 GS9000406    GS90015-02   |
-|                                                                          |-|
-|                                                   6264          6264     | |
-|                                     AC91106_VER1.7_4.IC63                | |
-|C                                                                         | |
-|U                                    AC91106_VER1.7_3.IC62                | |
-|S                                                                         | |
-|T                                                    MR91042-07-R66_6.IC95| |
-|O                                       62256                             |-|
-|M                                                                          |
-|7                                       62256    GS9000406    GS90015-02   |
-|2                     ADC0804                                              |
-|P                                                               6264       |
-|I                                                                         |-|
-|N                                                GS90015-03     6264      | |
-|                                                                          | |
-|                                 M6295                     PR91042_5.IC91 | |
-|                                      MR91042-08_2.IC57                   | |
-|                                                                          | |
-|                                 M6295                                    | |
-|                PMI_PM7528            AC91106_VER1.2_1.IC56               | |
-|                                                     62256                | |
-|                                                     62256                | |
-| POWER     DSW1   DSW2            4MHz                                    |-|
-|---------------------------------------------------------------------------|
-Notes:
-      68000 clock       - 12.0MHz [24/2]
-      Jaleco Custom ICs -
-                         GS9001501 (QFP44)
-                         GS9000406 (QFP80, x2)
-                         GS90015-02 (QFP100, x2)
-                         GS90015-03 (QFP80)
-
-      ROMs -
-            PR88004Q_8.IC102      - 82S147 PROM
-            AC91106_VER1.2_7.IC99 - 27C010 EPROM
-            MR91042007-R66_6.IC95 - 4M mask ROM
-            AC91106_VER1.7_4.IC63 - 27C010 EPROM
-            AC91106_VER1.7_3.IC62 - 27C010 EPROM
-            MR91042-08_2.IC57     - 4M mask ROM
-            AC91106_VER1.2_1.IC56 - 27C4001 EPROM
-            PR91042_5.IC91        - 82S129 PROM
-
-
-(lower board)
-GP-9189
-EB91015-20038
-|---------------------------------------------------------------------------|
-|MR91042-01-R60_1.IC1                                                       |
-|        MR91042-02-R61_2.IC2    62256                             62256    |
-|                                       GS90015-06    GS90015-06            |
-|MR91042-03-R62_3.IC5                                                       |
-|        MR91042-04-R63_4.IC6    62256                             62256    |
-|                                                                           |
-|MR91042-05-R64_5.IC11                                                      |
-|        MR91042-06-R65_6.IC12   62256                             62256   |-|
-|                                       GS90015-03    GS90015-03           | |
-|                                                                          | |
-|                                62256                             62256   | |
-|                                                                          | |
-|                                                                          | |
-|                        GS90015-08                                        | |
-|                                               GS90015-07   GS90015-10    |-|
-|                                                                           |
-|                                                                           |
-| CH9072-5_11.IC33                                                          |
-|              GS90015-11              CH9072-6_12.IC35                     |
-|                                      CH9072-4_13.IC39                    |-|
-|                                                                          | |
-|                                                                          | |
-|                          MR90015-35-W33_14.IC54                          | |
-|                                                                          | |
-|      CH9072-8_15.IC59    MR90015-35-W33_17.IC67                          | |
-|                                                    GS90015-09            | |
-|              PR88004W_16.IC66                                            | |
-|                                                                          | |
-|                                                      6116                | |
-| POWER                                                6116                |-|
-|---------------------------------------------------------------------------|
-Notes:
-      Jaleco Custom ICs -
-                         GS90015-03 (QFP80, x2)
-                         GS90015-06 (QFP100, x2)
-                         GS90015-07 (QFP64)
-                         GS90015-08 (QFP64)
-                         GS90015-09 (QFP64)
-                         GS90015-10 (QFP64)
-                         GS90015-11 (QFP100)
-
-      ROMs -
-            PR88004W_16.IC66      - 82S135 PROM
-            CH9072-4_13.IC39      - Atmel AT27HC642R
-            CH9072-5_11.IC33      - Atmel AT27HC642R
-            CH9072-6_12.IC35      - Atmel AT27HC642R
-            CH9072-8_15.IC59      - Atmel AT27HC642R
-            MR90015-35-W33_14.IC54- 4M mask ROM  \
-            MR90015-35-W33_17.IC67- 4M mask ROM  / Contain same data
-            MR91042-01-R60_1.IC1  - 4M mask ROM
-            MR91042-02-R61_2.IC2  - 4M mask ROM
-            MR91042-03-R62_3.IC5  - 4M mask ROM
-            MR91042-04-R63_4.IC6  - 4M mask ROM
-            MR91042-05-R64_5.IC11 - 4M mask ROM
-            MR91042-06-R65_6.IC12 - 4M mask ROM
-
-***************************************************************************/
-
-ROM_START( armchmp2 )
-	ROM_REGION( 0x080000, "maincpu", 0 )        /* Main CPU Code */
-	ROM_LOAD16_BYTE( "4_ver_2.7.ic63", 0x000000, 0x020000, CRC(e0cec032) SHA1(743b022b6de3efb045c4f1cca49caed0259ccfff) ) // same rom contents as ver 2.6, it appears to be correct;
-	ROM_LOAD16_BYTE( "3_ver_2.7.ic62", 0x000001, 0x020000, CRC(44186a37) SHA1(d21fbba11e9c82f48de2a699011ca7f3b90061ba) ) // modifications are in data area, where upper bytes are 00
-
-	ROM_REGION( 0x080000, "scroll0", 0 ) /* Scroll 0 */
-	ROM_LOAD( "mr91042-07-r66_6.ic95",  0x000000, 0x080000, CRC(d1be8699) SHA1(67563761f95892b08c7113ab1c52ab5aa7118fb8) )
-
-//  ROM_REGION( 0x080000, "scroll1", 0 ) /* Scroll 1 */
-//  UNUSED
-
-	ROM_REGION( 0x020000, "scroll2", 0 ) /* Scroll 2 */
-	ROM_LOAD( "ac91106_ver1.2_7.ic99", 0x000000, 0x020000, CRC(09755aef) SHA1(39c901fb9408a0ba488f0112d7f48b929b092e3b) )
-
-	ROM_REGION( 0x300000, "sprites", 0 ) /* Sprites */
-	ROM_LOAD16_BYTE( "mr91042-01-r60_1.ic1",  0x000000, 0x080000, CRC(fdfe6951) SHA1(ba6c5cd5d16fdca6f131302b19e621f8abe8136a) )
-	ROM_LOAD16_BYTE( "mr91042-02-r61_2.ic2",  0x000001, 0x080000, CRC(2e6c8b30) SHA1(70503fec251606b37fea2c7f91e682aece252035) )
-	ROM_LOAD16_BYTE( "mr91042-03-r62_3.ic5",  0x100000, 0x080000, CRC(07ba6d3a) SHA1(9c58e3a1931b593448c53a59e7f5b9aaac40ff88) )
-	ROM_LOAD16_BYTE( "mr91042-04-r63_4.ic6",  0x100001, 0x080000, CRC(f37cb12c) SHA1(282ebbd795284d7efa335b797ca1eedc1110e9da) )
-	ROM_LOAD16_BYTE( "mr91042-05-r64_5.ic11", 0x200000, 0x080000, CRC(7a3bb52d) SHA1(7f9d1dad4c89e6b55415b082363bc261115e9f96) )
-	ROM_LOAD16_BYTE( "mr91042-06-r65_6.ic12", 0x200001, 0x080000, CRC(5312a4f2) SHA1(4dcd2839bb5acccecf1eb6c0e19e877a0cff6875) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "mr91042-08_2.ic57",     0x000000, 0x080000, CRC(dc015f6c) SHA1(9d0677c50a25be1d11d43e54dbf3005f18b79b66) )
-	ROM_RELOAD(                        0x080000, 0x080000 )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "ac-91106v2.0_1.ic56", 0x000000, 0x080000, CRC(0ff5cbcf) SHA1(25ef8d67749ca78afc4c13a31b3f7a87284947c1) )
-	ROM_RELOAD(                        0x080000, 0x080000 )
-
-	ROM_REGION( 0x80000, "user2", 0 )       /* ? Unused ROMs ? */
-	ROM_LOAD( "ch9072-4_13.ic39", 0x000000, 0x002000, CRC(b45b4dc0) SHA1(b9fae0c9ac2d40f0a202c538d866d5f2941ba8dd) )
-	ROM_LOAD( "ch9072-5_11.ic33", 0x000000, 0x002000, CRC(e122916b) SHA1(86d5ecc7ecc6f175ecb28459697ef33e1ee06860) )
-	ROM_LOAD( "ch9072-6_12.ic35", 0x000000, 0x002000, CRC(05d95bf7) SHA1(78181cf71f22c090a1e62823a43757353a9ef6ab) )
-	ROM_LOAD( "ch9072-8_15.ic59", 0x000000, 0x002000, CRC(6bf52596) SHA1(bf4e7e7df3daae4aa6a441b58b15a435aa45630e) )
-
-	ROM_LOAD( "mr90015-35-w33_17.ic67", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-	ROM_LOAD( "mr90015-35-w33_14.ic54", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-
-	ROM_LOAD( "pr88004q_8.ic102", 0x000000, 0x000200, CRC(9327dc37) SHA1(cfe7b144cdcd76170d47f1c4e0f72b6d4fca0c8d) )
-	ROM_LOAD( "pr88004w_16.ic66", 0x000000, 0x000100, CRC(3d648467) SHA1(bf8dbaa2176c801f7370313425c87f0eefe8a3a4) )
-
-	ROM_LOAD( "pr91042_5.ic91", 0x000000, 0x000100, CRC(e71de4aa) SHA1(d06e5a35ad2127df2d6328cce153073380ee7819) )
-ROM_END
-
-ROM_START( armchmp2o2 )
-	ROM_REGION( 0x080000, "maincpu", 0 )        /* Main CPU Code */
-	ROM_LOAD16_BYTE( "ac-91106v2.6_4.ic63", 0x000000, 0x020000, CRC(e0cec032) SHA1(743b022b6de3efb045c4f1cca49caed0259ccfff) )
-	ROM_LOAD16_BYTE( "ac-91106v2.6_3.ic62", 0x000001, 0x020000, CRC(5de6da19) SHA1(1f46056596924789394ad2d99ec2d7fcb7845d3c) )
-
-	ROM_REGION( 0x080000, "scroll0", 0 ) /* Scroll 0 */
-	ROM_LOAD( "mr91042-07-r66_6.ic95",  0x000000, 0x080000, CRC(d1be8699) SHA1(67563761f95892b08c7113ab1c52ab5aa7118fb8) )
-
-//  ROM_REGION( 0x080000, "scroll1", 0 ) /* Scroll 1 */
-//  UNUSED
-
-	ROM_REGION( 0x020000, "scroll2", 0 ) /* Scroll 2 */
-	ROM_LOAD( "ac91106_ver1.2_7.ic99", 0x000000, 0x020000, CRC(09755aef) SHA1(39c901fb9408a0ba488f0112d7f48b929b092e3b) )
-
-	ROM_REGION( 0x300000, "sprites", 0 ) /* Sprites */
-	ROM_LOAD16_BYTE( "mr91042-01-r60_1.ic1",  0x000000, 0x080000, CRC(fdfe6951) SHA1(ba6c5cd5d16fdca6f131302b19e621f8abe8136a) )
-	ROM_LOAD16_BYTE( "mr91042-02-r61_2.ic2",  0x000001, 0x080000, CRC(2e6c8b30) SHA1(70503fec251606b37fea2c7f91e682aece252035) )
-	ROM_LOAD16_BYTE( "mr91042-03-r62_3.ic5",  0x100000, 0x080000, CRC(07ba6d3a) SHA1(9c58e3a1931b593448c53a59e7f5b9aaac40ff88) )
-	ROM_LOAD16_BYTE( "mr91042-04-r63_4.ic6",  0x100001, 0x080000, CRC(f37cb12c) SHA1(282ebbd795284d7efa335b797ca1eedc1110e9da) )
-	ROM_LOAD16_BYTE( "mr91042-05-r64_5.ic11", 0x200000, 0x080000, CRC(7a3bb52d) SHA1(7f9d1dad4c89e6b55415b082363bc261115e9f96) )
-	ROM_LOAD16_BYTE( "mr91042-06-r65_6.ic12", 0x200001, 0x080000, CRC(5312a4f2) SHA1(4dcd2839bb5acccecf1eb6c0e19e877a0cff6875) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "mr91042-08_2.ic57",     0x000000, 0x080000, CRC(dc015f6c) SHA1(9d0677c50a25be1d11d43e54dbf3005f18b79b66) )
-	ROM_RELOAD(                        0x080000, 0x080000 )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "ac-91106v2.0_1.ic56", 0x000000, 0x080000, CRC(0ff5cbcf) SHA1(25ef8d67749ca78afc4c13a31b3f7a87284947c1) )
-	ROM_RELOAD(                        0x080000, 0x080000 )
-
-	ROM_REGION( 0x80000, "user2", 0 )       /* ? Unused ROMs ? */
-	ROM_LOAD( "ch9072-4_13.ic39", 0x000000, 0x002000, CRC(b45b4dc0) SHA1(b9fae0c9ac2d40f0a202c538d866d5f2941ba8dd) )
-	ROM_LOAD( "ch9072-5_11.ic33", 0x000000, 0x002000, CRC(e122916b) SHA1(86d5ecc7ecc6f175ecb28459697ef33e1ee06860) )
-	ROM_LOAD( "ch9072-6_12.ic35", 0x000000, 0x002000, CRC(05d95bf7) SHA1(78181cf71f22c090a1e62823a43757353a9ef6ab) )
-	ROM_LOAD( "ch9072-8_15.ic59", 0x000000, 0x002000, CRC(6bf52596) SHA1(bf4e7e7df3daae4aa6a441b58b15a435aa45630e) )
-
-	ROM_LOAD( "mr90015-35-w33_17.ic67", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-	ROM_LOAD( "mr90015-35-w33_14.ic54", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-
-	ROM_LOAD( "pr88004q_8.ic102", 0x000000, 0x000200, CRC(9327dc37) SHA1(cfe7b144cdcd76170d47f1c4e0f72b6d4fca0c8d) )
-	ROM_LOAD( "pr88004w_16.ic66", 0x000000, 0x000100, CRC(3d648467) SHA1(bf8dbaa2176c801f7370313425c87f0eefe8a3a4) )
-
-	ROM_LOAD( "pr91042_5.ic91", 0x000000, 0x000100, CRC(e71de4aa) SHA1(d06e5a35ad2127df2d6328cce153073380ee7819) )
-ROM_END
-
-ROM_START( armchmp2o )
-	ROM_REGION( 0x080000, "maincpu", 0 )        /* Main CPU Code */
-	ROM_LOAD16_BYTE( "ac91106_ver1.7_4.ic63", 0x000000, 0x020000, CRC(aaa11bc7) SHA1(ac6186f45a006074d3a86d7437c5a3411bf27188) )
-	ROM_LOAD16_BYTE( "ac91106_ver1.7_3.ic62", 0x000001, 0x020000, CRC(a7965879) SHA1(556fecd6ea0f977b718d132c4180bb2160b9da7e) )
-
-	ROM_REGION( 0x080000, "scroll0", 0 ) /* Scroll 0 */
-	ROM_LOAD( "mr91042-07-r66_6.ic95",  0x000000, 0x080000, CRC(d1be8699) SHA1(67563761f95892b08c7113ab1c52ab5aa7118fb8) )
-
-//  ROM_REGION( 0x080000, "scroll1", 0 ) /* Scroll 1 */
-//  UNUSED
-
-	ROM_REGION( 0x020000, "scroll2", 0 ) /* Scroll 2 */
-	ROM_LOAD( "ac91106_ver1.2_7.ic99", 0x000000, 0x020000, CRC(09755aef) SHA1(39c901fb9408a0ba488f0112d7f48b929b092e3b) )
-
-	ROM_REGION( 0x300000, "sprites", 0 ) /* Sprites */
-	ROM_LOAD16_BYTE( "mr91042-01-r60_1.ic1",  0x000000, 0x080000, CRC(fdfe6951) SHA1(ba6c5cd5d16fdca6f131302b19e621f8abe8136a) )
-	ROM_LOAD16_BYTE( "mr91042-02-r61_2.ic2",  0x000001, 0x080000, CRC(2e6c8b30) SHA1(70503fec251606b37fea2c7f91e682aece252035) )
-	ROM_LOAD16_BYTE( "mr91042-03-r62_3.ic5",  0x100000, 0x080000, CRC(07ba6d3a) SHA1(9c58e3a1931b593448c53a59e7f5b9aaac40ff88) )
-	ROM_LOAD16_BYTE( "mr91042-04-r63_4.ic6",  0x100001, 0x080000, CRC(f37cb12c) SHA1(282ebbd795284d7efa335b797ca1eedc1110e9da) )
-	ROM_LOAD16_BYTE( "mr91042-05-r64_5.ic11", 0x200000, 0x080000, CRC(7a3bb52d) SHA1(7f9d1dad4c89e6b55415b082363bc261115e9f96) )
-	ROM_LOAD16_BYTE( "mr91042-06-r65_6.ic12", 0x200001, 0x080000, CRC(5312a4f2) SHA1(4dcd2839bb5acccecf1eb6c0e19e877a0cff6875) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "mr91042-08_2.ic57",     0x000000, 0x080000, CRC(dc015f6c) SHA1(9d0677c50a25be1d11d43e54dbf3005f18b79b66) )
-	ROM_RELOAD(                        0x080000, 0x080000 )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* Samples (4x40000) */
-	ROM_LOAD( "ac91106_ver1.2_1.ic56", 0x000000, 0x080000, CRC(48208b69) SHA1(5dfcc7744f7cdd0326886a4a841755ab6ec5482b) )
-	ROM_RELOAD(                        0x080000, 0x080000 )
-
-	ROM_REGION( 0x80000, "user2", 0 )       /* ? Unused ROMs ? */
-	ROM_LOAD( "ch9072-4_13.ic39", 0x000000, 0x002000, CRC(b45b4dc0) SHA1(b9fae0c9ac2d40f0a202c538d866d5f2941ba8dd) )
-	ROM_LOAD( "ch9072-5_11.ic33", 0x000000, 0x002000, CRC(e122916b) SHA1(86d5ecc7ecc6f175ecb28459697ef33e1ee06860) )
-	ROM_LOAD( "ch9072-6_12.ic35", 0x000000, 0x002000, CRC(05d95bf7) SHA1(78181cf71f22c090a1e62823a43757353a9ef6ab) )
-	ROM_LOAD( "ch9072-8_15.ic59", 0x000000, 0x002000, CRC(6bf52596) SHA1(bf4e7e7df3daae4aa6a441b58b15a435aa45630e) )
-
-	ROM_LOAD( "mr90015-35-w33_17.ic67", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-	ROM_LOAD( "mr90015-35-w33_14.ic54", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-
-	ROM_LOAD( "pr88004q_8.ic102", 0x000000, 0x000200, CRC(9327dc37) SHA1(cfe7b144cdcd76170d47f1c4e0f72b6d4fca0c8d) )
-	ROM_LOAD( "pr88004w_16.ic66", 0x000000, 0x000100, CRC(3d648467) SHA1(bf8dbaa2176c801f7370313425c87f0eefe8a3a4) )
-
-	ROM_LOAD( "pr91042_5.ic91", 0x000000, 0x000100, CRC(e71de4aa) SHA1(d06e5a35ad2127df2d6328cce153073380ee7819) )
-ROM_END
-
-/***************************************************************************
-
-                                Captain Flag
-
-**************************************************************************/
-
-ROM_START( captflag )
-	ROM_REGION( 0x80000, "maincpu", 0 )        /* Main CPU Code */
-	ROM_LOAD16_BYTE( "cf-92128a_3_ver1.4.ic40", 0x000000, 0x020000, CRC(e62af6ae) SHA1(38eb581def468860e4705f25088550799303a9aa) )
-	ROM_LOAD16_BYTE( "cf-92128_4_ver1.4.ic46",  0x000001, 0x020000, CRC(e773f87f) SHA1(cf9d72b0df256b69b96f1cd6b5f86282801873e3) )
-
-	ROM_REGION( 0x80000, "scroll0", 0 ) /* Scroll 0 */
-	ROM_LOAD( "mr92027-11_w89.ic54", 0x000000, 0x080000, CRC(d34cae3c) SHA1(622ad4645df12d34e55bbfb7194508957bb2198b) ) // 5 on the PCB
-
-//  ROM_REGION( 0x080000, "scroll1", 0 ) /* Scroll 1 */
-//  UNUSED
-
-	ROM_REGION( 0x20000, "scroll2", 0 ) /* Scroll 2 */
-	ROM_LOAD( "cf-92128_6.ic55", 0x000000, 0x020000, CRC(12debfc2) SHA1(f28d3f63a3c8965fcd838eedad4ef3682a28da0d) ) // 6 on the PCB
-
-	ROM_REGION( 0x400000, "sprites", 0 ) /* Sprites */
-	ROM_LOAD16_BYTE( "gp-9189_1.ic1",       0x000000, 0x080000, CRC(03d69f0f) SHA1(97a0552d94ca1e9c76896903c02c3f005752e5db) )
-	ROM_LOAD16_BYTE( "gp-9189_2.ic2",       0x000001, 0x080000, CRC(fbfba282) SHA1(65735d8f6abdb5816b8b26ce2969959a0f2e2c7d) )
-	ROM_LOAD16_BYTE( "mr92027-08_w85.ic5",  0x100000, 0x080000, CRC(1e02abff) SHA1(497aeab40c778ef6d8a76b005530fea5f4f68ab8) )
-	ROM_LOAD16_BYTE( "mr92027-04_w86.ic6",  0x100001, 0x080000, CRC(89dfe3e8) SHA1(f069cf55d08e5901699e7ee7c6b2b5cdba031cf2) )
-	ROM_LOAD16_BYTE( "gp-9189_5.ic11",      0x200000, 0x080000, CRC(edc33285) SHA1(a23f9a1877108eaaef6d467ae21433fe5f24261f) )
-	ROM_LOAD16_BYTE( "gp-9189_6.ic12",      0x200001, 0x080000, CRC(99b8f6d8) SHA1(e01c85861d5a131bb31908460cb3bc9eb0ea045c) )
-	ROM_LOAD16_BYTE( "mr92027-07_w87.ic15", 0x300000, 0x080000, CRC(07e49754) SHA1(c9fcd394badba104508e4c5ec5cf5513ecb2d69e) )
-	ROM_LOAD16_BYTE( "mr92027-08_w88.ic16", 0x300001, 0x080000, CRC(fb080dd6) SHA1(49eceba8cdce76dec3f6a85327135125bb0910f0) )
-
-	ROM_REGION( 0x80000, "user2", 0 )       /* ? Unused ROMs ? */
-	ROM_LOAD( "pr91042.ic86",        0x000000, 0x000100, CRC(e71de4aa) SHA1(d06e5a35ad2127df2d6328cce153073380ee7819) ) // FIXED BITS (00000xxx0000xxxx)
-	ROM_LOAD( "pr88004q.ic88",       0x000000, 0x000200, CRC(9327dc37) SHA1(cfe7b144cdcd76170d47f1c4e0f72b6d4fca0c8d) ) // FIXED BITS (1xxxxxxx1111x1xx)
-	ROM_LOAD( "pr92027a.ic16",       0x000000, 0x000020, CRC(bee7adc6) SHA1(cd11a3dae0317d06a69b5707a653b8997c1eb97f) ) // FIXED BITS (0000000000000xxx), 1xxx0 = 0x00
-	ROM_LOAD( "pr92027a.ic17",       0x000000, 0x000020, CRC(bee7adc6) SHA1(cd11a3dae0317d06a69b5707a653b8997c1eb97f) ) // ""
-	ROM_LOAD( "pr92027b.ic32.bin",   0x000000, 0x000113, CRC(483f4fb5) SHA1(84bb0300a106261634c921a37858482d3233c05a) )
-	ROM_LOAD( "pr92027b.ic32.jedec", 0x000000, 0x000bd0, CRC(f0ed1845) SHA1(203438fdee05810b2265624e1a1fdd55d360f833) )
-	ROM_LOAD( "pr92027b.ic36.bin",   0x000000, 0x000113, CRC(483f4fb5) SHA1(84bb0300a106261634c921a37858482d3233c05a) )
-	ROM_LOAD( "pr92027b.ic36.jedec", 0x000000, 0x000bd0, CRC(f0ed1845) SHA1(203438fdee05810b2265624e1a1fdd55d360f833) )
-
-	ROM_LOAD( "ch9072_4.ic39",       0x000000, 0x002000, CRC(b45b4dc0) SHA1(b9fae0c9ac2d40f0a202c538d866d5f2941ba8dd) ) // FIXED BITS (0000000x), 1ST AND 2ND HALF IDENTICAL
-	ROM_LOAD( "ch9072_5.ic33",       0x000000, 0x001000, CRC(a8025dc1) SHA1(c9bb7ea59bba3041c687b449ff1560d7d1ce2ec9) ) // FIXED BITS (xxxx0xxx)
-	ROM_LOAD( "ch9072_6.ic35",       0x000000, 0x001000, CRC(5cc9c561) SHA1(10866fd0707498fe4d4415bf755c07b55af4ae18) )
-	ROM_LOAD( "ch9072_8.ic59",       0x000000, 0x001000, CRC(6c99523b) SHA1(cc00b326b69a97b5bd2e2d741ab41692a14eae35) ) // FIXED BITS (0xxx0xxx)
-
-	ROM_LOAD( "mr90015-35_w33.ic54", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) ) // not dumped yet (taken from the other games)
-	ROM_LOAD( "mr90015-35_w33.ic67", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) ) // ""
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* Samples (8x20000) */
-	ROM_LOAD( "mr92027-10_w27.ic19", 0x000000, 0x100000, CRC(04bd729e) SHA1(92bcedf16554f33cc3d0dbdd8807b0e2fafe5d7c) ) // 2 on the PCB
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* Samples (8x20000) */
-	ROM_LOAD( "mr92027-09_w26.ic18", 0x000000, 0x100000, CRC(3aaa332a) SHA1(6c19364069e0b077a07ac4f9c4b0cf0c0985a42a) ) // 1 on the PCB
-ROM_END
-
-void captflag_state::init_captflag()
-{
-	m_oki1_bank->configure_entries(0, 0x100000 / 0x20000, memregion("oki1")->base(), 0x20000);
-	m_oki2_bank->configure_entries(0, 0x100000 / 0x20000, memregion("oki2")->base(), 0x20000);
-}
-
-
-/***************************************************************************
-
-                         Vs. Super Captain Flag
-
- PCB IDs:
-
- CF-93153 EB93041-20091
-
- CF-92128B EB92027-20060
-
- GP-9189 EB90015-20038
-
-**************************************************************************/
-
-ROM_START( vscaptfl )
-	ROM_REGION( 0x80000, "maincpu", 0 )        /* Main CPU Code */
-	ROM_LOAD16_BYTE( "vs.s.c.f-cf-92128-3_ver.1.3.ic40", 0x000000, 0x040000, CRC(a52595af) SHA1(6a7b3172c256015a0b46394dc1d0bf92ef826a90) )
-	ROM_LOAD16_BYTE( "vs.s.c.f-cf-92128-4_ver.1.3.ic46", 0x000001, 0x040000, CRC(1d015f55) SHA1(652d8c3e5c71ae149ff95a008a15d067bcdd8293) )
-
-	ROM_REGION( 0x80000, "scroll0", 0 ) /* Scroll 0 */
-	ROM_LOAD( "mr92027_11.ic54", 0x000000, 0x080000, CRC(d34cae3c) SHA1(622ad4645df12d34e55bbfb7194508957bb2198b) )
-
-//  ROM_REGION( 0x080000, "scroll1", 0 ) /* Scroll 1 */
-//  UNUSED
-
-	ROM_REGION( 0x20000, "scroll2", 0 ) /* Scroll 2 */
-	ROM_LOAD( "vs-cf92128_6_ver1.0.ic55", 0x000000, 0x020000, CRC(55c8db06) SHA1(499c0de2202e873eeca98c2f0932569f3df01a21) )
-
-	ROM_REGION( 0x500000, "sprites", 0 ) /* Sprites */
-	ROM_LOAD16_BYTE( "mr92027_01.ic1",           0x000000, 0x080000, CRC(03d69f0f) SHA1(97a0552d94ca1e9c76896903c02c3f005752e5db) )
-	ROM_LOAD16_BYTE( "mr92027_02.ic2",           0x000001, 0x080000, CRC(fbfba282) SHA1(65735d8f6abdb5816b8b26ce2969959a0f2e2c7d) )
-	ROM_LOAD16_BYTE( "mr92027_03.ic5",           0x100000, 0x080000, CRC(1e02abff) SHA1(497aeab40c778ef6d8a76b005530fea5f4f68ab8) )
-	ROM_LOAD16_BYTE( "mr92027_04.ic6",           0x100001, 0x080000, CRC(89dfe3e8) SHA1(f069cf55d08e5901699e7ee7c6b2b5cdba031cf2) )
-	ROM_LOAD16_BYTE( "mr92027_05.ic11",          0x200000, 0x080000, CRC(edc33285) SHA1(a23f9a1877108eaaef6d467ae21433fe5f24261f) )
-	ROM_LOAD16_BYTE( "mr92027_06.ic12",          0x200001, 0x080000, CRC(99b8f6d8) SHA1(e01c85861d5a131bb31908460cb3bc9eb0ea045c) )
-	ROM_LOAD16_BYTE( "vs-gp9189_7_ver1.0.ic15",  0x300000, 0x080000, CRC(63459f84) SHA1(1956ceab56f7266329b4e276876363f2334f0287) )
-	ROM_LOAD16_BYTE( "vs-gp9189_8_ver1.0.ic16",  0x300001, 0x080000, CRC(ed40b351) SHA1(9fca6ec3f364ee80729d59e617b2341d39326f2b) )
-	ROM_LOAD16_BYTE( "vs-gp9189_9_ver1.0.ic21",  0x400000, 0x080000, CRC(eb20c7d4) SHA1(64956c672fec1e96a97e25144b351bc24c32aa4f) )
-	ROM_LOAD16_BYTE( "vs-gp9189_10_ver1.0.ic22", 0x400001, 0x080000, CRC(5e24c3b0) SHA1(714418a48d66782798abfb9394b3a527e63ad9c5) )
-
-	ROM_REGION( 0x80000, "user2", 0 )       /* Unused ROMs */
-	// Located on CF-93153
-	ROM_LOAD( "pr92027b.ic31",       0x000000, 0x000117, CRC(bebd5200) SHA1(6ce078daf44b7839536dae0e6539388228f4ff47) )
-	ROM_LOAD( "pr92027b.ic34",       0x000000, 0x000117, CRC(bebd5200) SHA1(6ce078daf44b7839536dae0e6539388228f4ff47) )
-	// Located on CF-92128B
-	ROM_LOAD( "pr88004q.ic88",       0x000000, 0x000200, CRC(9327dc37) SHA1(cfe7b144cdcd76170d47f1c4e0f72b6d4fca0c8d) )
-	ROM_LOAD( "pr91042.ic86",        0x000000, 0x000100, CRC(e71de4aa) SHA1(d06e5a35ad2127df2d6328cce153073380ee7819) )
-	ROM_LOAD( "pr92027a.ic16",       0x000000, 0x000020, CRC(bee7adc6) SHA1(cd11a3dae0317d06a69b5707a653b8997c1eb97f) )
-	ROM_LOAD( "pr92027a.ic17",       0x000000, 0x000020, CRC(bee7adc6) SHA1(cd11a3dae0317d06a69b5707a653b8997c1eb97f) )
-	// Located on CF-92128B
-	ROM_LOAD( "ch9072_4.ic48",       0x000000, 0x002000, CRC(b45b4dc0) SHA1(b9fae0c9ac2d40f0a202c538d866d5f2941ba8dd) )
-	ROM_LOAD( "ch9072_5.ic36",       0x000000, 0x002000, CRC(e122916b) SHA1(86d5ecc7ecc6f175ecb28459697ef33e1ee06860) )
-	ROM_LOAD( "ch9072_6.ic35",       0x000000, 0x002000, CRC(05d95bf7) SHA1(78181cf71f22c090a1e62823a43757353a9ef6ab) )
-	ROM_LOAD( "ch9072_8.ic65",       0x000000, 0x002000, CRC(6bf52596) SHA1(bf4e7e7df3daae4aa6a441b58b15a435aa45630e) )
-	ROM_LOAD( "mr90015-35_w33.ic54", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-	ROM_LOAD( "mr90015-35_w33.ic67", 0x000000, 0x080000, CRC(9d428fb7) SHA1(02f72938d73db932bd217620a175a05215f6016a) )
-
-	ROM_REGION( 0x100000, "oki1", 0 )       /* Samples (8x20000) */
-	ROM_LOAD( "vs-cf92128_2.ic19", 0x000000, 0x100000, CRC(7629969c) SHA1(94b391f230f5b5f9498bc967beaed1e6c95a028f) )
-
-	ROM_REGION( 0x100000, "oki2", 0 )       /* Samples (8x20000) */
-	ROM_LOAD( "vs-cf92128_1.ic18", 0x000000, 0x100000, CRC(50c99a47) SHA1(c0c4ac24eebdd462176a5cbaf9e33918c6d54e10) )
-ROM_END
-
-void captflag_state::init_vscaptfl()
-{
-	address_space &space = m_maincpu->space(AS_PROGRAM);
-
-	space.unmap_readwrite(0x100044, 0x100045);
-	space.unmap_readwrite(0x100048, 0x100049);
-
-	space.install_read_port(0x100044, 0x100045, "Buttons2");
-	space.install_write_handler(0x10004c, 0x10004d, write16s_delegate(*this, FUNC(captflag_state::motor_command_left_w)));
-	space.install_write_handler(0x100050, 0x100051, write16s_delegate(*this, FUNC(captflag_state::motor_command_right_w)));
-
-	init_captflag();
-}
-
+} // anonymous namespace
 
 /***************************************************************************
 
@@ -4130,25 +3128,15 @@ void captflag_state::init_vscaptfl()
 
 ***************************************************************************/
 
-GAMEL( 1989, bigrun,    0,        bigrun_d65006,   bigrun,   cischeat_state,  init_bigrun,   ROT0,   "Jaleco", "Big Run (11th Rallye version, Europe?)", MACHINE_NODEVICE_LAN, layout_cischeat ) // there's a 13th Rallye version (1991) (only on the SNES? Could just be updated title, 1989 -> 11th Paris-Dakar ...)
-GAMEL( 1989, bigrunu,   bigrun,   bigrun_d65006,   bigrun,   cischeat_state,  init_bigrun,   ROT0,   "Jaleco", "Big Run (11th Rallye version, US?)",     MACHINE_NODEVICE_LAN, layout_cischeat )
+GAMEL( 1989, bigrun,    0,        bigrun_d65006, bigrun,   cischeat_state,  init_bigrun,   ROT0,   "Jaleco", "Big Run (11th Rallye version, Europe?)", MACHINE_NODEVICE_LAN, layout_cischeat ) // there's a 13th Rallye version (1991) (only on the SNES? Could just be updated title, 1989 -> 11th Paris-Dakar ...)
+GAMEL( 1989, bigrunu,   bigrun,   bigrun_d65006, bigrun,   cischeat_state,  init_bigrun,   ROT0,   "Jaleco", "Big Run (11th Rallye version, US?)",     MACHINE_NODEVICE_LAN, layout_cischeat )
 
-GAMEL( 1990, cischeat,  0,        cischeat_gs88000, cischeat, cischeat_state,  init_cischeat, ROT0,   "Jaleco", "Cisco Heat",                             MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_cischeat )
+GAMEL( 1990, cischeat,  0,        cischeat,      cischeat, cischeat_state,  init_cischeat, ROT0,   "Jaleco", "Cisco Heat",                             MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_cischeat )
 
-GAMEL( 1992, f1gpstar,  0,        f1gpstar, f1gpstar, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "Grand Prix Star (ver 4.0)",              MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
-GAMEL( 1991, f1gpstar3, f1gpstar, f1gpstar, f1gpstar, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "Grand Prix Star (ver 3.0)",              MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
-GAMEL( 1991, f1gpstar2, f1gpstar, f1gpstar, f1gpstar, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "Grand Prix Star (ver 2.0)",              MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
+GAMEL( 1992, f1gpstar,  0,        f1gpstar,      f1gpstar, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "Grand Prix Star (ver 4.0)",              MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
+GAMEL( 1991, f1gpstar3, f1gpstar, f1gpstar,      f1gpstar, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "Grand Prix Star (ver 3.0)",              MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
+GAMEL( 1991, f1gpstar2, f1gpstar, f1gpstar,      f1gpstar, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "Grand Prix Star (ver 2.0)",              MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
 
-GAME(  1992, armchmp2,  0,        armchmp2, armchmp2, armchamp2_state, empty_init,    ROT270, "Jaleco", "Arm Champs II (ver 2.7)",                MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING | MACHINE_MECHANICAL )
-GAME(  1992, armchmp2o2,armchmp2, armchmp2, armchmp2, armchamp2_state, empty_init,    ROT270, "Jaleco", "Arm Champs II (ver 2.6)",                MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING | MACHINE_MECHANICAL )
-GAME(  1992, armchmp2o, armchmp2, armchmp2, armchmp2, armchamp2_state, empty_init,    ROT270, "Jaleco", "Arm Champs II (ver 1.7)",                MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING | MACHINE_MECHANICAL )
+GAME(  1992, wildplt,   0,        wildplt,       wildplt,  wildplt_state,   init_f1gpstar, ROT0,   "Jaleco", "Wild Pilot",                             MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // busted timings
 
-GAME(  1992, wildplt,   0,        wildplt,  wildplt,  wildplt_state,   init_f1gpstar, ROT0,   "Jaleco", "Wild Pilot",                             MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING ) // busted timings
-
-GAMEL( 1993, f1gpstr2,  0,        f1gpstr2, f1gpstr2, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "F-1 Grand Prix Star II",                 MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
-
-GAME(  1993, captflag,  0,        captflag, captflag, captflag_state,  init_captflag, ROT270, "Jaleco", "Captain Flag (Japan)",                   MACHINE_IMPERFECT_GRAPHICS )
-GAME(  1993, vscaptfl,  0,        captflag, vscaptfl, captflag_state,  init_vscaptfl, ROT270, "Jaleco", "Vs. Super Captain Flag (Japan)",         MACHINE_IMPERFECT_GRAPHICS )
-
-GAME(  1994, scudhamm,  0,        scudhamm, scudhamm, cischeat_state,  empty_init,    ROT270, "Jaleco", "Scud Hammer (ver 1.4)",                  MACHINE_IMPERFECT_GRAPHICS ) // version on program ROMS' labels
-GAME(  1994, scudhamma, scudhamm, scudhamm, scudhamma,cischeat_state,  empty_init,    ROT270, "Jaleco", "Scud Hammer (older version?)",           MACHINE_IMPERFECT_GRAPHICS ) // maybe older cause it has one less dip listed?
+GAMEL( 1993, f1gpstr2,  0,        f1gpstr2,      f1gpstr2, cischeat_state,  init_f1gpstar, ROT0,   "Jaleco", "F-1 Grand Prix Star II",                 MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN, layout_f1gpstar )
