+++ src/mame/jaleco/ms1_tmap.h
@@ -28,32 +28,32 @@ class megasys1_tilemap_device : public device_t, public device_gfx_interface
 public:
 	// construction/destruction
 	template <typename T>
-	megasys1_tilemap_device(const machine_config &mconfig, const char *tag, device_t *owner, T &&palette_tag, uint16_t colorbase)
-		: megasys1_tilemap_device(mconfig, tag, owner, (uint32_t)0)
+	megasys1_tilemap_device(const machine_config &mconfig, const char *tag, device_t *owner, T &&palette_tag, u16 colorbase)
+		: megasys1_tilemap_device(mconfig, tag, owner, (u32)0)
 	{
 		set_palette(std::forward<T>(palette_tag));
 		set_colorbase(colorbase);
 	}
 
-	megasys1_tilemap_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock);
+	megasys1_tilemap_device(const machine_config &mconfig, const char *tag, device_t *owner, u32 clock);
 
 	// configuration
 	void set_8x8_scroll_factor(int scroll_factor) { m_8x8_scroll_factor = scroll_factor; }
 	void set_16x16_scroll_factor(int scroll_factor) { m_16x16_scroll_factor = scroll_factor; }
 	void set_bits_per_color_code(int bits) { m_bits_per_color_code = bits; }
-	void set_colorbase(uint16_t colorbase) { m_colorbase = colorbase; }
+	void set_colorbase(u16 colorbase) { m_colorbase = colorbase; }
 	template <typename T> void set_screen_tag(T &&tag) { m_screen.set_tag(std::forward<T>(tag)); }
 
 	// memory handlers
-	void write(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
-	uint16_t scroll_r(offs_t offset);
-	void scroll_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+	void write(offs_t offset, u16 data, u16 mem_mask = ~0);
+	u16 scroll_r(offs_t offset);
+	void scroll_w(offs_t offset, u16 data, u16 mem_mask = ~0);
 
 	// drawing and layer control
-	void draw(screen_device &screen, bitmap_ind16 &dest, const rectangle &cliprect, uint32_t flags, uint8_t priority = 0, uint8_t priority_mask = 0xff);
+	void draw(screen_device &screen, bitmap_ind16 &dest, const rectangle &cliprect, u32 flags, u8 priority = 0, u8 priority_mask = 0xff);
 	void enable(bool enable);
-	void set_flip(uint32_t attributes);
-	void set_tilebank(uint8_t bank);
+	void set_flip(u32 attributes);
+	void set_tilebank(u8 bank);
 
 protected:
 	// device-level overrides
@@ -63,23 +63,23 @@ protected:
 
 private:
 	// shared memory finder
-	required_shared_ptr<uint16_t> m_scrollram;
+	required_shared_ptr<u16> m_scrollram;
 	optional_device<screen_device> m_screen;
 
 	// configuration
 	int m_8x8_scroll_factor;
 	int m_16x16_scroll_factor;
 	int m_bits_per_color_code;
-	uint16_t m_colorbase;
+	u16 m_colorbase;
 
 	// decoding info
 	DECLARE_GFXDECODE_MEMBER(gfxinfo);
 
 	// internal state
-	uint16_t m_scrollx = 0;
-	uint16_t m_scrolly = 0;
-	uint16_t m_scroll_flag = 0;
-	uint16_t m_tile_bank = 0;
+	u16 m_scrollx = 0;
+	u16 m_scrolly = 0;
+	u16 m_scroll_flag = 0;
+	u32 m_tile_bank = 0;
 	tilemap_t *m_tmap = nullptr;
 	tilemap_t *m_tilemap[2][4]{};
